Support Microsoft::WRL::ComPtr as a receiver of base::Bind{Once,Repeating}
After this CL, we can use ComPtr as the |this| pointer of base::BindOnce
and base::BindRepeating.

ComPtr doesn't have `operator*` nor `operator->*` that are needed to run
a method pointer. So we need a special handling for it to use it as a
receiver on base::Bind.

We currently use scoped_refptr to store COM instance on base::Bind, which
happens to work, but that looks fragile. This CL is a preparation to
convert these scoped_refptr to ComPtr.

Change-Id: I1fe31e23ac627050860fe2c81ead6807a17bb939
Reviewed-on: https://chromium-review.googlesource.com/1145153
Commit-Queue: Taiju Tsuiki <tzik@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#577470}

CrOS-Libchrome-Original-Commit: c44f810cbe81532a80b76259d267f063a253278f