Android: Removed all Params, Progress from AsyncTask
All changes (except for the changes to AsyncTask itself) are entirely
generated with Android Studio's refactor mechanism and git cl format.

This is in an effort to reduce the unused surface of AsyncTask that
doesn't match the ideal future API. Ideally, we will move to something
very similiar to TaskScheduler, which has no concept of "Progress
updates", and forces the usages to call Bind for the parameters.

For parameters now, we use constructor arguments, which provides the
same concurrency guarantees that passing in params via Params would.
This is closer to the ideal because it supports any arbitrary
parameters, not just many of 1 type.

TBR=smaier  #For mechanical refactoring

Bug: 843745
Cq-Include-Trybots: master.tryserver.chromium.android:android_cronet_tester;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: If25df5d7276e504983bdc71e6e07993cb044ed04
Reviewed-on: https://chromium-review.googlesource.com/1150692
Commit-Queue: Sam Maier <smaier@chromium.org>
Reviewed-by: Sam Maier <smaier@chromium.org>
Reviewed-by: agrieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#580549}

CrOS-Libchrome-Original-Commit: 1a696109129c8382b0f756cfc365b2edf41f820c