[Base] Use background mode for ThreadPriority::BACKGROUND threads (behind feature) (reland).
Relanding a CL that was reverted because of a variable that was
unused in non-DCHECK builds and caused a compile failure.
Original CL: Reviewed-on: https://chromium-review.googlesource.com/1171482

This CL adds a feature to use THREAD_MODE_BACKGROUND_BEGIN instead of
THREAD_PRIORITY_LOWEST for ThreadPriority::BACKGROUND threads. This
lowers the disk and network I/O priority of the thread in addition to
the CPU scheduling priority. MSDN recommends using this setting for
threads that perform background work.
https://docs.microsoft.com/en-us/windows/desktop/api/processthreadsapi/nf-processthreadsapi-setthreadpriority

TBR=gab@chromium.org

Bug: 872820
Change-Id: I212fb77bb035088595b4944ce145a74d333c38eb
Reviewed-on: https://chromium-review.googlesource.com/1185529
Commit-Queue: François Doray <fdoray@chromium.org>
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#585272}

CrOS-Libchrome-Original-Commit: d88c3ec18b44afe5a8f4a5b147e7e15dead47426