Make TimeDelta TriviallyCopyable for std::atomic.
We're having some awkwardness with working around the issue with stuff
like std::atomic<int64_t> and conversions back and forth. It would be
preferable to use std::atomic<TimeDelta> directly.

Removes the workaround added for bug 635974, since that was a bug in
vs 2015 and we use clang now.

Also fixes a couple of lint issues in time.h.

Bug: 635974, 851959, 761570
Change-Id: I4683f960b0c348748c5f0aaf222da4dda40256ec
Reviewed-on: https://chromium-review.googlesource.com/1184781
Commit-Queue: Max Morin <maxmorin@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#586219}

CrOS-Libchrome-Original-Commit: db5f2b6e74d6d2e2a9c6f530199067c1d6eff7b6