Add a sampling profiler to chrome tracing.
This CL is plumbing the base::StackProfiling to allow collecting
stack sample at a low rate on the browser process.
An event is triggered to the trace with the collected stackframe.

The goal of this code is to get low-rate stack samples in the field
for the slow-report to help the investigation of stalled tasks.


This is a variant of:
  https://chromium-review.googlesource.com/c/chromium/src/+/1188502

R=ssid@chromium.org, oysteine@chromium.org

Change-Id: I150c17df51914f4535d4e7d6ec086a2e13c17cff
Reviewed-on: https://chromium-review.googlesource.com/1191745
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Mike Wittman <wittman@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Reviewed-by: Alexei Filippov <alph@chromium.org>
Reviewed-by: Siddhartha S <ssid@chromium.org>
Cr-Commit-Position: refs/heads/master@{#588554}

CrOS-Libchrome-Original-Commit: 0ea3570a2882d86f47bf2cebf3e85a9fa16cd7a5