TaskScheduler: Replace "background" with "best effort" when related to tasks/sequences.
This CL does not update the word "background" when referring to
the "background" thread priority.

This CL does not change the name of UMA histograms, because this
would affect externally observable behavior. We will make this change
in a separate CL.

This CL does not change the name of the --disable-background-tasks
flag because it would require updating tests outside of //base that
use this flag. We will make this change in a separate CL.

This is a no-op CL.

Bug: 863341
Change-Id: I8fad34a56aa134cf5837d92c7794e1461fa541d0
Reviewed-on: https://chromium-review.googlesource.com/1213542
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#590699}

CrOS-Libchrome-Original-Commit: 79bf554374fb4b76d09f86e674da8747c08b219c