[mojo-core] Reduce lingering message queue costs
Incoming messages are stored in a std::vector which we never shrink when
popping off elements. This means its capacity may be allowed to grow
indefinitely, which is bad.

This CL adds a check when popping messages off the queue,
calling shrink_to_fit() if the operation results in the queue crossing
an arbitrary size threshold (multiples of 512 elements).

Bug: 888665
Change-Id: Ie2b89a2104bcd6d3eece29c339e297cdadbd837b
Reviewed-on: https://chromium-review.googlesource.com/1241561
Commit-Queue: Ken Rockot <rockot@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#594104}

CrOS-Libchrome-Original-Commit: 7fe00cd1270a395d3f1d66b14e41595a432d6ddc