Remove fetching lsb contents from chrome
Partners will frequently share debug logs to provide context when filing
a bug or diagnosing a problem. Feedback reports have this information,
but it's not stored in any of the files that are packaged up with debug
logs.

This commit follows the CL below, which makes debugd responsible for
fetching the lsb release contents:
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1217083

BUG=chromium:880851
TEST=Manually checked chrome://system to ensure that lsb release
contents are present even after removing them from the chrome source.
Also checked that chrome://system loads properly and with correct output
when both debugd and chrome are reporting the lsb release contents.

Change-Id: I4651c5168c26ac696d5799c44126e5dcd6b4ce19
Reviewed-on: https://chromium-review.googlesource.com/1217344
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Commit-Queue: Kartik Hegde <khegde@chromium.org>
Cr-Commit-Position: refs/heads/master@{#594453}

CrOS-Libchrome-Original-Commit: dcbbb2fda2c82e5b83bfb6a5a0177cfbb83f6b8b