Add helpers for byte[] <-> std::string.
Several places in the code convert between Java byte arrays and C++
std::string, and some had implemented a local helper function for this.
Add a generic helper for this to base instead, and convert existing
places that do this to use the helper.

Test: base_unittests
Change-Id: Ia56fb6333bd6c453207814b0bcb5ac58b5488c20
Reviewed-on: https://chromium-review.googlesource.com/1255342
Commit-Queue: Richard Coles <torne@chromium.org>
Reviewed-by: Pavel Yatsuk <pavely@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#595863}

CrOS-Libchrome-Original-Commit: cd5fa3ffc58ea937aaa55d0c500c2b81f70162a8