[Fuchsia] Fix a potential race condition in test_launcher.
The test launcher creates a /data directory for each of its
children. This is implemented as a directory with a unique number in
its name. Access to the number was not thread-safe, which this CL
fixes by using atomic operations to increase the number.

This is a potential fix for the content_unittests failure seen here:
https://ci.chromium.org/p/chromium/builders/luci.chromium.try/fuchsia_x64/117497

Change-Id: I42614b0de36e65ffb8bc811813132a8781036896
Reviewed-on: https://chromium-review.googlesource.com/1255926
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Fabrice de Gans-Riberi <fdegans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#595887}

CrOS-Libchrome-Original-Commit: 0b8758b800d61f67e9e0fb6690c7ddc9bb915ac0