Base: Replace CurrentProcessInfo::CreationTime() with Process::CreationTime().
The new API supports getting the creation time of any process (not just
the current process).

This will be used to get a (pid, creation time) pair identifying a
renderer process and sending it to the resource coordinator service
via Mojo. Ideally, we would send a process handle, but this is not
supported by Mojo, and if it was, it would have undesirable security
implications.

Bug: 885293
Change-Id: Ic3a24c93e5352767eb8e6fb025eeacdbc0d23109
Reviewed-on: https://chromium-review.googlesource.com/c/1273607
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#601423}

CrOS-Libchrome-Original-Commit: 7d131534d96bcd0344b5b3d44c737de178df8f15