[Password export][Android] Create file in native code
When Chrome exports passwords, it creates a temporary file, serializes
the passwords into it and shares it through an intent.

So far, creating the file happened from the Java code, while writing
to it from the native C++ backend. This required gathering and
reporting I/O errors from two places.

This CL moves the file creation next to the writing code in the native
C++ backend.

As a side effect to achieve the goal, this CL also adds a two-argument
callback for the communication across Java bindings, because
Callback.java only accepts a single argument. The added callback has
fixed argument types (and int and a string), because generalising
seems like a significant effort and there is no clear demand for
that. OTOH, the new callback ends up in base, next to the single-arg
callback, so that it is visible to developers in the future.

Bug: 891218
Change-Id: Ida685baa72456f8631356f219ea80ad7db195db5
Reviewed-on: https://chromium-review.googlesource.com/c/1256813
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Vaclav Brozek <vabr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#603376}

CrOS-Libchrome-Original-Commit: ddb5f9f89cb68eed764f8d0a7fca4af238a53d74