[TaskScheduler] Alive/dead instant events instead of begin/end
This is preferable because it avoids adding trace events while the
thread is idle (for which there is no precedent).

The delta between the instant events can be used to perform the same
analysis which could be performed previously (and these will be in
Slow Reports per being in an enabled category).

R=fdoray@chromium.org

Bug: 899897
Change-Id: I40d31df9f4db3ea8959f760323b7cd8f56299ce3
Reviewed-on: https://chromium-review.googlesource.com/c/1312928
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#605017}

CrOS-Libchrome-Original-Commit: 2f1e97157af097d62ac6084c7edc8f35a0cd2461