[TaskScheduler]: Update ScopedAllowBaseSyncPrimitives friend list.
Preparing work for migrating ThreadRestrictions::ScopedAllowWait to
ScopedAllowBaseSyncPrimitives.
This CL updates the friendlist for ScopedAllowBaseSyncPrimitives and
ScopedAllowBaseSyncPrimitivesWithoutBlockingScope.
Some cases might need to be adjusted if it was labeled wrong, but we can assume
that most cases are correct. This saves us from updating this file for each
cases.

Bug: 766678
Change-Id: Ie11bca133f7664355425cfaefcffb4acc0373731
Reviewed-on: https://chromium-review.googlesource.com/c/1324100
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#606250}

CrOS-Libchrome-Original-Commit: 1cad95562f0fbd28741173bc8e4ba82fb0eb8488