Fix InspectConstructionArguments with random min & max.
If the min/max parameters are really fubar, such as which might come
from a malicious IPC message, min & max can be limited only to then
be swapped and become invalid.  Do the swap first.

Also, remove duplicate bucket-limit condition and set the defined
maximum to be the lower of the two.

Bug: 905226
Change-Id: If73e97e43e93704f93456da6af0708422fa145a0
Reviewed-on: https://chromium-review.googlesource.com/c/1335731
Commit-Queue: Brian White <bcwhite@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#608398}

CrOS-Libchrome-Original-Commit: 7cfaffb1b47f922c8d51c26d2846692f85d9e608