Remove the PlatformSharedMemoryRegion permissions check on NaCl
To perform the permissions check on NaCl, we mmap() a shared memory region as
writable and check the return value to know if map failed. It's expected that
the call would fail if the region has read-only permissions.

Unfortunately, NaCl doesn't handle failure cases gracefully and crashes
instead of returning error value [1], [2].

This means, that an attempt to check permissions on a read-only shared memory
region always leads to a crash. This CL removes the check altogether to allow
the NaCl processes to use read-only shared memory.

[1] Posix: https://cs.chromium.org/chromium/src/native_client/src/trusted/service_runtime/sys_memory.c?rcl=f701a90597fc85979319447c0cd44c3b52201c78&l=868
[2] Win: https://cs.chromium.org/chromium/src/native_client/src/shared/imc/win/nacl_shm.cc?rcl=f701a90597fc85979319447c0cd44c3b52201c78&l=112

Bug: 901646
Change-Id: I1f0cee4f7edae0390701320f5061470d62240186
Reviewed-on: https://chromium-review.googlesource.com/c/1352312
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Alexandr Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#611630}

CrOS-Libchrome-Original-Commit: b76c606d8656f4373ac8591d6300a93ffc384154