RemoteMacViews: Break focus manager query across mojo
Break the function -[ChromeCommandDispatcherDelegate
eventHandledByExtensionCommand] into the parts that are to
run in the PWA process and the parts that are to run in the
browser process.

Rename the function to eventHandledByViewsFocusManager, as is
requested by the TODOs.

Fix-up the mojo support for ui::Accelerator, since it is used by this
method (support was mostly written, but was not actually included in
the build).

Bug: 895169
Change-Id: Ic9180ac55fcd7d0bce8e23e6d0c7a5e4be4cb8b0
Reviewed-on: https://chromium-review.googlesource.com/c/1358641
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#614068}

CrOS-Libchrome-Original-Commit: 642609c77909916568c41482e1ae56e1ca8db2ec