[mojo-core] Add Channel Message type validation
This effectively corrects a DCHECK in Message deserialization code by
aligning Channel::Message::Deserialize and the Message constructor on
their expectation of header sizes for different message types.

Deserialize was assuming that any type other than NORMAL must have a
legacy header, while the constructor assumes that any type other than
NORMAL_LEGACY must have a non-legacy header. In normal circumstances
this doesn't matter because Deserialize is only ever intentionally used
on NORMAL or NORMAL_LEGACY messages, but in the context of a fuzzer or a
misbehaving sender, it is possible for this mismatch to violate a
DCHECK (see the bug).

This changes Deserialize to match the constructor, so anything other
than NORMAL_LEGACY is treated as having a non-legacy header.

Bug: 912304
Change-Id: If5f35a7723004d79447831eeb80c3ed14e5977a4
Reviewed-on: https://chromium-review.googlesource.com/c/1363806
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#614128}

CrOS-Libchrome-Original-Commit: a1238b1b85d66e4df89305261f3e817aff3b03fb