TestBrowserThreadBundle to inherit from ScopedTaskEnvironment
This clarifies ownership of the ScopedTaskEnvironment (there is only one)
which makes it easier for the upcoming BrowserThreadScheduler work with
TestBrowserThreadBundle.  Because we need BrowserUIThreadScheduler and
ScopedTaskEnvironment to share a SequenceManager.

Most of this change is mechanical, the interesting changes are in
test_browser_thread_bundle.h, test_browser_thread_bundle.cc and
views_test_base.cc.

TBR=bartfab@chromium.org

Bug: 863341, 891670
Change-Id: I324665b19c6e89da42c05e7646ae41d9f1cd5e7b
Reviewed-on: https://chromium-review.googlesource.com/c/1352260
Reviewed-by: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#616622}

CrOS-Libchrome-Original-Commit: 1052bb0ca304bbdc91feacbeafe1acc9fe8fda76