Avoid instantiating SaturateFastAsmOp<> for types that do not use it.
Currently, this path is avoided via an |if (false) { ... }| in
the code above it when SaturateFastAsmOp is not supported for a
given pair of types. However, instantiating the template for 64-bit
sources will give a truncation warning for arm (at the line where
we do |int32_t src = value;|) under strict compiler modes (even
though the path can't be reached).

Avoid the error by conditionalizing its instatiation for cases where
it is actually usable.

BUG=879657

Change-Id: I40bee5a4686b89c1d35c7b93804b215afe25dd68
Reviewed-on: https://chromium-review.googlesource.com/c/1387958
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#618634}

CrOS-Libchrome-Original-Commit: f1e233b3fb8afdcd0cff83be8c20d891ba1d8271