Refactor WebThreadImpl, WebSubThread, and WebMainLoop.
WebThreadImpl is now merely a scoped object that binds a provided
SingleThreadTaskRunner to a WebThread::ID. It no longer subclasses
base::Thread. That functionality has been moved to WebSubThread, which is
effectively only used for the IO thread now.

This is the //web implementation of
https://chromium-review.googlesource.com/c/chromium/src/+/980793. See the
description of that CL for more details.

BUG=826465

Change-Id: I636051b154c4156bc32517d649081131f9bdd14d
Reviewed-on: https://chromium-review.googlesource.com/c/1383140
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#618761}

CrOS-Libchrome-Original-Commit: 752eeba72e3136a4d81cfbcff56d6dffd1338ac0