TaskScheduler: Add metrics to track cancelled delayed tasks
In order to determine the best approach to removing cancelled tasks from
the DelayedTaskManager, it is necessary to collect metrics regarding the
number of cancelled tasks that remain in the heap, as well as the
percentage tasks in the heap that are cancelled. Given these metrics,
we hope to deduce the number and percentage of cancelled tasks at which
the benefits of cleaning up cancelled tasks outweigh the cost.

Bug: 900191
Change-Id: Ibecf6aff5c5176deba0059607945e9b2c6dbfbf8
Reviewed-on: https://chromium-review.googlesource.com/c/1423159
Commit-Queue: Aditya Keerthi <adityakeerthi@google.com>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#626072}

CrOS-Libchrome-Original-Commit: 98dbee3a496427115cca76409e8ed40dd021a7d9