Mojo: Avoid pseudo handle weirdness in fuzzing
Windows reserves a few special handle values as "pseudo handles". These
can have surprising interactions with e.g. DuplicateHandle when the
source process is the calling process. These turn out to be problematic
in a fuzzing environment where our "remote" sending process is actually
the same as the receiving process and thus we may attempt to duplicate
fuzzed pseudo handle values with bad results.

This adds a simple check for decoded handle values to ensure that they
aren't one of the known pseudo handle values before attempting to
call DuplicateHandle on them. While this wouldn't be problematic when
the handles actually come from a remote process (DuplicateHandle will
fail), the presence of such handle values in a message indicates a
misbehaving sender and so we might as well reject the message anyway,
even a normal production setting.

Bug: 925651
Change-Id: Ia41676b505fb550f631338645f1366194d1a18d3
Reviewed-on: https://chromium-review.googlesource.com/c/1438814
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#626738}

CrOS-Libchrome-Original-Commit: 282d0372d1ce5d0795a62c3c30b49856d0cfea84