[TaskScheduler]: Unflake TaskSchedulerWorkerPoolBlockingTest.
Tests in TaskSchedulerWorkerPoolBlockingTest were assuming that |max_tasks_| is
increased to number of blocked tasks + initial capacity, which may never be true
when workers are idle. This causes flakes especially when
|MayBlockThresholdMicroseconds| == |BlockedWorkersPollMicroseconds|.

This CL removes the assertions and instead posts more tasks that wait for each other,
in order to make sure that |max_tasks_| is increased when necessary.

Bug: 874080
Change-Id: I24d4b5ad2e2b777528d6027ad355c794b8bef453
Reviewed-on: https://chromium-review.googlesource.com/c/1439842
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#627141}

CrOS-Libchrome-Original-Commit: 1b735bd628c43acb7241137bdf68b55558f852b1