Refuse to download new toolchain during gn gen
This change relies on the depot_tools change crrev.com/c/1439805
Chromium and v8 have rolled to a new enough version of depot_tools to
support this script and Angle always uses the latest version of
depot_tools so there shouldn't be any script-mismatch problems caused
by landing this.

The toolchain used to build Chrome is supposed to be updated during
"gclient runhooks" but currently if that doesn't happen (if
GYP_MSVS_VERSION is changed but "gclient runhooks" isn't run) then the
update will be tried during "gn gen". That works badly. It leads to
attempts to parse the update output as toolchain information, it can
lead to the update happening in parallel, and is generally a bad idea.

This change passes --no-download to
depot_tools\win_toolchain\get_toolchain_if_necessary.py when 'update'
is not the command being processed so that the depot_tools script will
no that it should only configure the toolchain, not try downloading it.
This also prevents attempts to update the toolchain when running
tools\win\setenv.bat

This was tested by forcing an updated to VS 2019 and confirming that
the new behavior is as desired (failing with an informative message
when appropriate, working when appropriate). This includes testing
that using the local toolchain (if installed) works with this change.
See crrev.com/c/1426397, the CL that was used for testing on the bots.

Both bugs suggest that our bots sometimes don't update the toolchain
when they are supposed to. If that is the case then this change will
help to flush out those bugs so that they can be fixed.

Bug: 584393,662325
Change-Id: I0ef671977ea59b4f3368827c5b5ffe50ed6602f7
Reviewed-on: https://chromium-review.googlesource.com/c/1444031
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Scott Graham <scottmg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#627513}

CrOS-Libchrome-Original-Commit: 060aa690a1abce245184e1f251e9414a0eb8637f