Get TypeId to work with the msvc compiler.
Last attempt to land a unittests that makes sure that TypeId works
across shared libraries failed on MSVC and was reverted:
http://crrev/c/1375293

This patch keeps the current code for Clang. For MSVC instead of
using the address of a templated static variable as unique id it makes
use of the limited functionality the typeid operator provides when
compiling on MSVC without RTTI.

The only downside of using typeid is that it is not a constexpr and
thus many of the TypeId member methods can not be either. We could
remove constexpr-ness on MSVC builds only but to keep the interface
consistent across configurations this patch removes it for all of them.
We might want to revisit this decision later.

Bug: 914651

Change-Id: I458f41074fb04a6e459810fbc8c6d99a8d25707d
Reviewed-on: https://chromium-review.googlesource.com/c/1375952
Commit-Queue: Carlos Caballero <carlscab@google.com>
Reviewed-by: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#628718}

CrOS-Libchrome-Original-Commit: 1c04255879f37c9baf3789a99dcf53de262b41da