[Promises] Introduce base::unique_any
base::unique_any is similar to std::any except it doesn't use exceptions or
RTTI which are banned in Chromium (std::any sort of works with RTTI disabled
but it's crippled because you can't use Any::type to inspect the type).

base::unique_any can hold any move or copy constructible type (std::any
can only hold copy constructible types).  As you might expect
base::unique_any doesn't have a copy constructor or a copy assignment
operator.  I chose not to implement base::any for now since the main
motivation for base::unique_any is to allow base::AbstractPromise
to be an ordinary type and to reduce promise template bloat. Still it
would be easy to implement base::any on top of this patch if we chose to.

Small objects are stored inline and larger ones with an auxiliary
allocation.

Bug: 906125
Change-Id: I16223d27aa46490a1e36266c59264ec24ddc5062
Reviewed-on: https://chromium-review.googlesource.com/c/1340802
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#629107}

CrOS-Libchrome-Original-Commit: 22e57499f6f2391e58ddab58db93736b034fa840