TaskScheduler: Change SchedulerWorkerPool::(ReenqueueSequence->GetWorkerPoolForTraits).
With this CL, when the pool wants to reenqueue a Sequence, it asks
TaskScheduler what pool is appropriate for the Sequence's traits
and it performs the reenqueue itself. Previously, it handed off
the Sequence to the TaskScheduler and TaskScheduler performed the
reenqueue.

This change will allow the pool to avoid a lock re-acquisition in
the case that the Sequence stays in the same pool.

Bug: 889029
Change-Id: I6bd4322d5e1eb4d220a80a708c0644dec2fe88af
Reviewed-on: https://chromium-review.googlesource.com/c/1450457
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#629221}

CrOS-Libchrome-Original-Commit: f48ece8245eb078a8d7e29f5f87463e068ebef70