TaskScheduler: Ensure that we always wake up the correct number of workers.
This CL introduces SchedulerWorkerPoolImpl::WakeUpWorkersLockRequired(),
which wakes up the appropriate number of workers to run Sequences
in the PriorityQueue, given the current |max_tasks_|. It does not
wake up workers for Sequences that are in the PriorityQueue but
already had a worker woken up for them.

This method is reused when a task is posted, when a sequence is
moved between pools and when the |max_tasks_| is updated due to
a ScopedBlockingCall.

Bug: 889029, 757897
Change-Id: I35ae855bc56d42ecae22e171787ea8ffc50d2920
Reviewed-on: https://chromium-review.googlesource.com/c/1450287
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#630763}

CrOS-Libchrome-Original-Commit: 69f66f05be3167319c445bf5e46c7da30da47141