metrics: Convert FieldTrialList to the new shared memory API
This is a refactoring CL that switches all usages of obsolete base::SharedMemory
and base::SharedMemoryHandle classes in FieldTrialList to the new shared memory
API.

No behavioral change is expected.

Field trials are written in a shared memory region once in the browser process
and shared with all other processes as read-only. Thus, the
base::ReadOnlySharedMemoryRegion class is used.

FieldTrialList requires an access to internals of a shared memory region to
implement passing a descriptor/handle to a child process at the moment of
process creation. ReadOnlyShmRegion supports only self-destructive way of
getting a platform handle to the shared memory. FieldTrialList requires an
exception to this rule, so it's added as a friend of ReadOnlyShmRegion.

Some random files started missing "base/files/file_util.h" include that used
to be included transitively. This CL adds an explicit include to those files.

TBR: sky@chromium.org
Bug: 920183
Change-Id: I7a7e276c5afa631d56ceb7b2afa6f6597099f176
Reviewed-on: https://chromium-review.googlesource.com/c/1402880
Auto-Submit: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#631312}

CrOS-Libchrome-Original-Commit: 027ca3d3f91e39618221f16b5e91983d72c3b40f