Experiment: Schedule font unique name lookup table build at startup
Modify DWriteFontLookupTableBuilder to support off-main-thread
scheduling of the font unique name lookup table construction.

For now, schedule this as a USER_BLOCKING-prioritized startup task in a
separate SequencedTaskRunner. Synchronize using a WaitableEvent that the
DWriteFontProxyImpl can wait on when it receives an incoming Mojo
request from a renderer for this table.

Once the UpdatePriority() API for SequencedTaskRunners becomes
available, downgrade the priority to BEST_EFFORT, then bump it once the
renderer requests the table. This TODO is tracked in [1].

Background for font unique name local lookups in the design review doc
in https://crbug.com/828317. See also startup job scheduling discussion
in [2].

[1] https://crbug.com/931366
[2] https://groups.google.com/a/chromium.org/d/msg/chromium-dev/iDClDAsZnrE/bb5eZXkUFAAJ

Bug: 889864
Change-Id: I8e19d34520422fc9404ca01cb8357221d7a11495
Reviewed-on: https://chromium-review.googlesource.com/c/1459635
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#632008}

CrOS-Libchrome-Original-Commit: e7fcc1a51172ec06b5ceb6f9c4e30f79835256ae