TaskScheduler: Add TrackedRef::get().
This method gives access to the raw pointer stored in a TrackedRef.
It is useful to pass a raw pointer instead of copying a TrackedRef
in a scope where it is not possible for the source TrackedRef to be
deleted (e.g. in a method of a class that owns a const TrackedRef).

Bug: 889029
Change-Id: I89f062dd229fa0aac5787ee72e47f2dd82a9a0d8
Reviewed-on: https://chromium-review.googlesource.com/c/1464880
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#632241}

CrOS-Libchrome-Original-Commit: b8f593dc1f9ad90fae1b9c7641f0078fe980ee8c