dbus: Accept a{say} as an alternative to a{sv}
The Property of type map<string, vector<byte>> was created to read
BlueZ's property which unfortunately was weirdly encoded as a{sv} (array
of dictionary of string -> variant) instead of a{say} (array of dictionary
of string -> array of bytes). In the future NewBlue won't code this
D-Bus data the same weird way as BlueZ but will use a{say} instead so
Chrome's reader has to be able to accept both format for the transition
period.

BUG=chromium:927021

Change-Id: I68bf79afe1899695fc0c04e9e429c7fc71efda6d
Reviewed-on: https://chromium-review.googlesource.com/c/1450260
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>
Commit-Queue: Sonny Sasaka <sonnysasaka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#634791}

CrOS-Libchrome-Original-Commit: 737ad09f48db5b20153aba1c77eb2f558702eb9f