Add Location annotation ScopedBlockingCall
This CL is adding the FROM_HERE location of the call sites to each
ScopedBlockingCall. This modification will provide the
detailed location to the trace events and allow developpers do
understand cause of janks of long tasks.

This CL is trying to catch most of the call sites using this pattern:

  ^(\s*)(base::)?ScopedBlockingCall ([a-zA-Z0-9_]*)\((base::)?BlockingType

     -->

  \1\2ScopedBlockingCall \3\(FROM_HERE, \4BlockingType

Over every .cc, .h and .mm files under chromium.


TBR=gab@chromium.org (mechanical change for //base)

Bug: 934302
Change-Id: I9b7a573ab84d81cae1b8fc173b6e8ff96155ccb6
Reviewed-on: https://chromium-review.googlesource.com/c/1483834
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#635580}

CrOS-Libchrome-Original-Commit: 436d42211ddc1c6bd20fa715260e7e476e0242d8