Add Location annotation ScopedBlockingCall
This CL is adding the FROM_HERE location of the call sites to each
ScopedBlockingCall. This modification will provide the
detailed location to the trace events and allow developpers do
understand cause of janks of long tasks.

This CL is adding manually the cases not caught by the previous CL:
  https://chromium-review.googlesource.com/c/chromium/src/+/1483834


TBR=gab@chromium.org (skipping owners for side-effects after //base review)

Bug: 934302
Change-Id: I04dc84ddc0b8824f91d7409bb06c74397ec7bf32
Reviewed-on: https://chromium-review.googlesource.com/c/1489110
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#635701}

CrOS-Libchrome-Original-Commit: 95a01c2ab0d46b8513865425cc650d87ef0ff074