arm64/compiler Remove thinLTO cache clearing.
In https://chromium-review.googlesource.com/c/chromium/src/+/1481336 support was
added for clearing the ThinLTO cache in order to address a bug in LLVM when
disabling outlining. However this caused problems with the bots where linking
was happening at the same time. As such, remove this step and remove the ability
to change the whether outlining is enabled as a gn flag.

BUG=chromium:931297,chromium:935985

Change-Id: I53a33f0178346697842d3f14428d4a8b82dd2409
Reviewed-on: https://chromium-review.googlesource.com/c/1491435
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Ross McIlroy <rmcilroy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#636046}

CrOS-Libchrome-Original-Commit: a9c3f5e58265c6be2c53d73ac9924a6d54e8d4cf