android: crazy_linker: Remove Breakpad Guard region.
This CL removes an odd but obsolete feature to the Chromium linker,
which added a 16 MiB guard region when loading a given native library.

This was done a long time ago when our build system used to process
'regular' shared libraries generated by the linker, in order to compress
relocations inside it using a custom scheme. Said relocations were
later uncompressed / processed at load time through a static C++
initializer.

The main drawback of this scheme was that it modified the
correspondance between virtual addresses in the ELF files, and their
actual mapped location. This created all kind of chaos for tools like
Breakpad, and the guard region was added to mitigate the issue.

This custom compression scheme has been completely discarded, in favor
of adding the necessary machinery to the linker itself, in order to
directly generate shared libraries with the right compressed
relocations (which are of course supported by the crazy linker), without
any virtual / mapping address space discrepancies.

As such, this trick is no longer needed, simplifying our code.

This will also make it considerably easier to implementing
android_dlopen_ext() in the crazy linker.

BUG=504410,936001
R=pasko@chromium.org, agrieve@chromium.org, cjgrant@chromium.org

Change-Id: I5a38b952442160072f0e24f48289f41f0269cdc6
Reviewed-on: https://chromium-review.googlesource.com/c/1491271
Reviewed-by: David Turner <digit@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: David Turner <digit@chromium.org>
Cr-Commit-Position: refs/heads/master@{#636071}

CrOS-Libchrome-Original-Commit: 6f3ac52946fe86ebcca7d934b05b15bf92e8dd62