Add Location annotation ScopedBlockingCall
This CL is adding the FROM_HERE location of the call sites to each
ScopedBlockingCall. This modification will provide the
detailed location to the trace events and allow developpers do
understand cause of janks of long tasks.

This CL is adding manually the cases not caught by the previous CL:
  https://chromium-review.googlesource.com/c/chromium/src/+/1483834


TBR=gab@chromium.org

Bug: 934302
Change-Id: Id2b959ef007fd7fec6d5b7cf8373f93e816c8195
Reviewed-on: https://chromium-review.googlesource.com/c/1491896
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#636086}

CrOS-Libchrome-Original-Commit: 3d58bbd040296878c231e43e0386d400b3af3b00