Move Chrome trace argument filter to services/tracing
This CL is split from:
https://chromium-review.googlesource.com/c/chromium/src/+/1491032
This does not move the webview whitelists into services/tracing.

The argument filter is set by chrome delegate in each process. But,
tracing service runs as a sandboxed process, where the filter is not
set. In Perfetto backend json convertor filters arguments for all events
in the service process. So, move the filter to services/ and ensure it
is always used.
In short term, webview will continue using the filter as it is till we
agree on how to properly serialize the filter for Perfetto bcakend.

BUG=925148
TBR=avi@chromium.org

Change-Id: Ifaa491cbaf215f718a534063d39cbd771261cb0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1506659
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Commit-Queue: ssid <ssid@chromium.org>
Cr-Commit-Position: refs/heads/master@{#638398}

CrOS-Libchrome-Original-Commit: 518d025c85b2004025019e1a5ac29305d2b95b1f