Clean up Jelly Bean API compatibility code.
This CL removes the special cases for Jelly Bean in the
ApiCompatibilityUtils in //base. After that it turned out that all the
APIs could be inlined.

One API had to be special cased, since it was more than a single line,
namely the isDeviceProvisioned(Context) method. It had null-checks for
the Context and ContentResolver, and they were kept to ensure this CL
has no functional changes.

Some API callers were using the result of !inlineMethod(...), and the
method itself was checking == against something, and for those the
== was flipped to != and the negation of the call was removed
after inlining.

The method getUriForImageCaptureFile(File) had some documentation, and
that was moved to the only method it ended up calling, since that was
lacking documentation.

Bug: 893231,923477
Change-Id: I43420a38f6fe191b2c489f2f1bec99160bbcc718
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1496522
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#638432}

CrOS-Libchrome-Original-Commit: 861b89541841e36a936ade98869a31218bac3e79