SequenceManager: Record asynchronous call stack into a crash key
SequenceManager currently records crash keys for one level of
asynchronous call stack information for tasks (i.e., the FROM_HERE of
the posting location) in the form of function and file name strings.
This information is useful for tracking down the source of crashing
tasks, but the cost of recording these strings is not insignificant
(~1.5% of overall renderer CPU time on Windows according to sampling
profiler data).

This patch adds an alternate mechanism for recording this information in
the form of hexadecimal code addresses. This form is cheaper to encode
and more compact: we can now fit two levels of asynchronous stack frames
in 64 characters. The downside is that the data now needs to be
symbolized. This can be down with crsym as described at [1].

We still leave the old crash keys in place in order to have reliable
data while this new alternative is being proven.

Note that we use a hand-rolled function to write out the hex addresses
because it is faster than snprintf. The full results are in [2][3], but
to summarize this function is 81% faster on Android than the current
code, compared to 33% for snprintf.

Bug : 897751

[1] https://chromium.googlesource.com/chromium/src/+/master/docs/debugging_with_crash_keys.md
[2] https://docs.google.com/spreadsheets/d/1YxuWobFQaMh-XkkwiCJFVVqnpLT6vcBdMTQ3Ntszf3M/edit#gid=0
[3] https://chromium-review.googlesource.com/c/chromium/src/+/1523094

Change-Id: I6f43c6d2a57ed3d2fc500cd7b051621aeb499590
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1518576
Commit-Queue: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640822}

CrOS-Libchrome-Original-Commit: f9431a203b1ddd5c98d62bc72a3abfe448c4c582