Add "v8.wasm" trace event category
This category exists since a long time, but it was never listed in the
builtin categories. In the past it showed up eventually anyway, but
this seems to have changed.
In any case, listing it here provides a better user experience because
the category will just always be there.

R=ahaas@chromium.org, oysteine@chromium.org

Change-Id: I156b00667e2575459c28f412f6921e8d3b5f6a61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1528535
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Commit-Queue: Clemens Hammacher <clemensh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642099}

CrOS-Libchrome-Original-Commit: 0b691aae54efc55ac98738e484429137ee7df59a