TaskScheduler: Remove LazyStalenessPolicy from IntrusiveHeap
LazyStalenessPolicy was initially introduced to IntrusiveHeap in order
to address the issue of cancelled tasks remaining on the
DelayedTaskHeap. In practice, this approach found 0 cancelled tasks on
the DelayedTaskHeap in over 99% of cases. This could indicate a flawed
strategy in detecting cancelled tasks, or that there is not a
significant number of cancelled tasks remaining on the DelayedTaskHeap.
In either case, the LazyStalenessPolicy provides no benefit to
TaskScheduler.

Bug: 900191
Change-Id: I8ac94fdb214d822c607e9ce3751b706620f34d91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1521453
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Alex Clarke <alexclarke@chromium.org>
Commit-Queue: Aditya Keerthi <adityakeerthi@google.com>
Cr-Commit-Position: refs/heads/master@{#642526}

CrOS-Libchrome-Original-Commit: 5064860ccf5ed15eab9aed04763e7426b0b41e0c