Introduce mojo::core::ChannelMac.
This is a new Mojo Channel implementation that uses Mach IPC as the
transport mechanism. Because Mach IPC is simplex, the channel
implementation performs an internal setup handshake to establish bi-
directional communication with its peer:

- If the channel is created with a Mach send right, it first creates a
  new Mach receive right. The channel then transmits a new send right
  for its receive right to its peer using the send right that the
  channel was created with.
- If the channel is created with a Mach receive right, it starts by
  waiting for a message. The first message received is the channel-
  internal handshake message, which contains the send right of its peer.

Unlike other Channel implementations, ChannelMac reads and writes in a
message-oriented manner, rather than a stream-oriented one. If the
message length is greater than the size of the send/receive buffer used
by the channel, the message data are transmitted using Mach out-of-line
memory. The size of the buffer is the system page size, which has been
empirically determined to be efficient for most workloads.

ChannelMac is also capable of transferring both Mach port rights (send
and receive) as well as POSIX file descriptors directly within the
message.

This new implementation is gated behind a feature flag, which is
currently disabled. Minor modifications to ChannelPosix's Mach support
are added to enable switching at run-time.

Bug: 932175
Change-Id: I5d3a004de1e869f1896a99ca93df6207d4d18e7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1531521
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643541}

CrOS-Libchrome-Original-Commit: 9f707e5e04598d8303fa99ca29eb507c839767d8