ui/gl: Fix synchronization bug for release fences with SurfaceControl.
The change in [1] made an incorrect assumption that
GLImage::BindTexImage will be issued each time the associated texture is
bound for a write so issuing a service wait could be deferred until then.
The call is actually made only once at initialization and the state is
cached as Texture::BOUND on the image. As a result we don't wait on the
release fence received from the framework when reusing buffers.

[1]: https://chromium-review.googlesource.com/c/chromium/src/+/1526638

R=piman@chromium.org

Bug: 945443
Change-Id: I870adfb66f2935ff77c41ea5c714e1458aaf6716
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1538562
Auto-Submit: Khushal <khushalsagar@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644036}

CrOS-Libchrome-Original-Commit: 9d0e9a2122eda570db0a35ffcd48afeb0182c692