[Sampling profiler] Second reland of "Move context frame recording outside the native unwinders"
Allows the native unwinders to be used to unwind from arbitrary frames,
not just from the context frame. This is required to support restarting
native frame unwinding after v8 unwinding.

This is a reland of 8967586b85fb58ee3001f278894ae1e0614d7b44 with a fix
for an MSAN crash.

Bug: 909957, 948302
Change-Id: If0172f8ffac4f056817d32de0c6ef27d403843b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1540285
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#646333}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549684
Cr-Commit-Position: refs/heads/master@{#647023}

CrOS-Libchrome-Original-Commit: ad37bd1edc7f65bb93b5c1f862c983d10846819a