Allow extensions to be managed from multiple sources by policy.
Instead of always picking the highest priority source for the extension
policies merge all valid values from all sources if the new policy is
set and enabled.

This is a temporary solution for the most affected policies while the
long term solution is designed.

BUG=907795
TEST=components_unittests, manually

Change-Id: I224a2427c2447f228d60a1eeb721159a9b66cc05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1538404
Commit-Queue: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647707}

CrOS-Libchrome-Original-Commit: eb2225b2d0c8417a2c769b3a32c1bb608a15d878