[base] Introduce base::Value ExtractKey and ExtractPath
This change introduces base::Value's ExtractKey and ExtractPath to
provide an API that can transfer ownership of values out of an existing
dictionary. This change is motivated by the fact that this was possible
in the old Remove*() API, but wasn't possible with RemoveKey and RemovePath.

Bug: 646113
Change-Id: I8de4956b78652f990ecbc3f20aa757bcd6ad1dd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1550471
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648125}

CrOS-Libchrome-Original-Commit: c209c7d766e286fc06162e935d656bda6e76ea92