[mojo-core] Add the concept of port slots
Ports are the two-way relocatable messaging endpoints used as a backing
for Mojo message pipe handles. Each port is essentially a message-
receiving queue, and the mutual independence of each port's queue is
fundamental to a port's ability to be relocated efficiently and thus be
suitable to support message pipes.

This CL introduces the concept of "slots" on an individual port,
effectively allowing for messages in the port's queue to be routed to
different consumers referencing the port. Messages are now sent on a
specific slot by the sender, and received on a corresponding peer slot
on the peer port. If multiple messages are received for different slots
on a port, they still must be read (and are still only exposed) in
sequential order.

This is a precursor to supporting shared message ordering across
multiple independent message pipes in Mojo, a feature which will
ultimately replace the bindings concept of "associated interfaces".

See bug for design doc link.

Bug: 941809
Change-Id: I84e6f35831d918178e9eb2cc32c8701d1808e4e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1538778
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#648299}

CrOS-Libchrome-Original-Commit: d3b3ab90a1f2f625a64c55e1c55c5b342b84b09f