[mojo-core] Introduce PortAttachment
This is a simple refactoring CL which replaces the list of PortName
values in a UserMessageEvent with a list of PortAttachment structures.
This new structure type is a PortName and an optional SlotId, allowing
for internal Mojo serialization to signal that a serialized message pipe
handle should be encoded as a spliced-in slot on the transmission pipe
rather than as an actual port transfer.

No functional changes are made here, this is just groundwork to reduce
churn in a follow-up CL.

See bug for design doc link.

Bug: 941809
Change-Id: Id1c33a1fd81cf3879c350a3a241ef0791ef60ad3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1542304
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648417}

CrOS-Libchrome-Original-Commit: d6d91a56b93b7b8e4cbb6769d3faad6123861131