Improve Python 3 compatibility of Android GN util scripts
Most of the changes are for converting relative imports to be absolute, due to new restrictions on relative imports in Python 3.
All callers where necessary were updated to work with the use of absolute imports.

The scripts still works with Python 2.
There are no intended behaviour changes.

Bug: 941669
Change-Id: I39a8642f6cc02b9ea79e3bb7c6da4002444f2dfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1536068
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Reviewed-by: David Turner <digit@chromium.org>
Commit-Queue: Raul Tambre <raul@tambre.ee>
Cr-Commit-Position: refs/heads/master@{#649114}

CrOS-Libchrome-Original-Commit: 33ead3324d266f7c78f77eab11c368be4d9f1a2c