Native Library improvements
This change adds LoadSystemLibrary to NativeLibrary to allow usage of
the NativeLibrary abstraction when loading system dlls and prevent dll
loading attacks. This change also moves ScopedNativeLibrary to base off
of ScopedGeneric so that the scoped class gets all the benefits of
ScopedGeneric without having to re-invent.

Bug: 1551709
Change-Id: I17a14678687bd7d167ab0dad1cff5dbce53c3313
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1560294
Commit-Queue: Cliff Smolinsky <cliffsmo@microsoft.com>
Reviewed-by: Antoine Labour <piman@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650522}

CrOS-Libchrome-Original-Commit: f395befb97ee4205d2c05b3bc712ce53c71a085e