Shared memory: Move DLOG to regular LOG in perms check.
This may shed some light on the cause of crashes in the attached bug.
The logging will only occur if a CHECK fails, so it does not introduce
extra noise to logcat. The bug is focusing on Android, but crashes
occasionally occur on other platforms as well.

Bug: 838365
Change-Id: Ie6e3ef4f04c04b7f196d7654dc2bd28c84628394
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1565003
Commit-Queue: Matthew Cary <mattcary@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Auto-Submit: Matthew Cary <mattcary@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650722}

CrOS-Libchrome-Original-Commit: fdbd5366335f587422fb03fc5d8eea268cd43990