Use fallocate() to reserve disk space for shared memory
Shared memory is allocated by creating a file in /dev/shm (by default)
that can run out of space. Chrome reserves the file size by calling
ftruncate() that creates a sparse file, so it succeeds even if
sufficient disk space is not available.

This could lead to a situation when a shared memory region is
successfully created but a subsequent access to a shared memory page
results in SIGBUS due to the disk being full.

Using fallocate() instead of ftruncate() eliminates this problem because
the former syscall always allocates space and it returns an error if the
disk is full.

Bug: 951431
Change-Id: Ief38fda6ca5941aa83a9a9261dc6bb78e029668b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1564005
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650763}

CrOS-Libchrome-Original-Commit: c5ba5ddfb9fa686486b1a8a436be08a7164a1631