build/locale_tool.py: Add update-expectations command.
This adds a new command 'update-expectations' that can
be used to update the translation_expectations.pyl file
used to drive the localized strings import/export process.

For now, this simply ensures that the list of supported
locales is properly written in the .pyl files. For example
when adding missing localized strings [1].

Another use case is, when in preparation of adding new
locales, one could use --add-locales, as in:

  build/locale_tool.py updadte-expectations --add-locales "aa zz"

To ensure these new locales are listed in the "languages"
dictionaries of all .pyl files.

+ Add a unit-tests command to run the internal unit-test
  suite. For now, it only covers a few new functions but
  this may grow over time.

BUG=944606
R=agrieve@chromium.org, govind@chromium.org, jshin@chromium.org

Change-Id: I8f00639e2ce30a235acf0ae58d8eb582ecab75f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1567670
Commit-Queue: David Turner <digit@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651006}

CrOS-Libchrome-Original-Commit: 6e6f63f4ef5d58d2d5c1ef9ddd467ce5bee495a0