Fix MOCK_TIME bug with racy cross thread immediate task posting
In https://crrev.com/c/1564008 we have a test that only cares about main
thread delayed tasks.  However it invokes an API which as a side effect
triggers cross thread PostTask to the main thread.  These tasks are
uninteresting except occasionally they happen in the window between
MockTimeDomain::DelayTillNextTask and
MockTimeDomain::MaybeFastForwardToNextTask.  Due to a bug in
MaybeFastForwardToNextTask it was treating this racy cross thread post
the same as running out of tasks and exiting early.

This is clearly undesirable and this patch fixes that behaviour.

Change-Id: I98a96c98c8ae56fd22bc2b9ea452d46565354127
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1565983
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651677}

CrOS-Libchrome-Original-Commit: e7eac012ddf1a834935bbb50a696a7f4d223767b