[Sampling profiler] Check post-unwind frame when testing unwinding
To properly test that a special case unwind has succeeded, we need to
check that we successfully unwound past the special case frame. This
change adds a frame before the frame under test and verifies that it
appears in the resulting stack.

There remains a fair amount of duplication in this code which will be
addressed in follow-on CLs.

Bug: 909957
Change-Id: I6b39102e7bf23496a4b849d1dbf1780db383aaa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1574218
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Charlie Andrews <charliea@chromium.org>
Auto-Submit: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652959}

CrOS-Libchrome-Original-Commit: b4437d4079e6cd841190105e5c8ec352b7b92212