[Sampling profiler] Split out test unwind scenario setup
Moves the setup for the test unwind scenario from the TargetThread into
a separate UnwindScenario class. Changes the unwind scenario to be
configured using callbacks rather than plain function pointers, to
provide the flexibility that will be required to inject a custom setup
function for V8 unwinds. While this approach has the disadvantage that
it loses the ability to know the exact stack layout, it still allows
frame existence and relative ordering to be checked which should be
sufficient to catch failures.

Bug: 909957
Change-Id: I5ddf18219becba922edcccde69bd09e1fbbd38f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1577849
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Charlie Andrews <charliea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653198}

CrOS-Libchrome-Original-Commit: 3f76345f0dd6d7670d37c84781f6cad27829ed40