[mojo] Remove pre-dispatch handler callbacks
This hook was added to help implement WebContentsFrameBindingSet a while
ago, but that can be implemented without this. In the interest of not
adding a similar hack to the new ReceiverSet, this removes the need for
it by lightly refactoring WebContentsFrameBindingSet. There are no
other uses of pre-dispatch handlers, so their support is removed from
BindingSet here.

Bug: 955171
Change-Id: I1803c2ed1e1aca010840c561047a082cde91669b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1581159
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654092}


CrOS-Libchrome-Original-Commit: 3b799c66ce48329823367d0a7e989d5c04c842bf