Use range-based for loop and more references
While investigating a UAF bug I noticed that a for loop was copying a
scoped_refptr, causing a wasteful increment/decrement of the reference
count. Fixing that was just a matter of making the range declaration a
reference, but while I was in the area I converted a couple of loops
from iterators to range-based loops, for modest readability and
efficiency gains.

Change-Id: I04534f7838c05afa6d38f3064c837bed46c60450
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1584565
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654144}


CrOS-Libchrome-Original-Commit: 0ead651a5b21f5d7e3472d543735998ce70bb65f