SequenceManager Settings to use Builder Pattern with option to log RunLoop Quit
This patch introduces the debug only option to log when the run loop quits which
is useful in diagnosing flaky tests.  Unfortunately the new bool in
SequenceManager::Settings is enough to make the clang plugin decide it needs an
out of line constructor / destructor. We can't do that because the existing
brace style initalizations no longer compile, so we refactor to use a builder
pattern.

Bug: 863341
Change-Id: I33fc28fa72fce4beead8a6c524d9f2aab93cf1f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1581802
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654838}


CrOS-Libchrome-Original-Commit: 5f3d461b6307a738f96430acee216ed7d996cc06