Support skipping XR tests at runtime
Adds the ability to skip XR browser tests at runtime if requirements for
a test are not met.

As a result, also makes the browser test launcher properly handle gtest
reporting that a test was skipped, as previously it was treated as
an unknown error.

Currently, the only requirement that can be checked is whether a DirectX
11.1-capable GPU is present on the machine, but additional requirements
can be easily added.

Bug: 951469
Change-Id: I81c4cccf83bed3c13ec1d82ad99a16c29fba7cb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1575030
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Bill Orr <billorr@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657747}


CrOS-Libchrome-Original-Commit: d5aeb6324bbbcb4362652392e418d63327db214a