Ensure extracted file suffix changes in local development
Due to changes in crrev/c/1501459 (version.gni), local builds now get a
full versionCode as it would if it were a release build. This broke
logic which assumed that all local builds would have a versionCode < 10.
That logic was being used to give new names to extracted files that had
changed so that they get re-loaded.

Bug: 958644
Test: Follow steps in crrbug.com/958644 using the IDS_SAVE_PASSWORD string
Change-Id: I43e88fdfbbba485426ee1ea38344a45ed5c4bccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1601680
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Drew Stonebraker <stonebraker@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657987}


CrOS-Libchrome-Original-Commit: 5591c5d36457c699e8d25b38cc744e1c012e9b52