Disable leaking mojo_unittests on LSAN.
See crrev.com/c/1590404 and the linked bug for context -- the goal is to
fix the broken memory-leak checking in Chromium's ASAN
(AddressSanitizer) trybots, which are supposed to also run LSAN
(LeakSanitizer).

Currently-leaking tests must be fixed or disabled before leak checking
can be enabled in the ASAN trybot.

WARNING: This will result in a loss of ASAN coverage for the affected
tests given that LSAN runs as part of the ASAN trybot.

If that's not acceptable, we can wait until the leaks are fixed before
fixing leak detection in trybots, but keep in mind that newly-introduced
leaks Chromium-wide won't be caught until then.

Bug: 969789
Change-Id: I543d75510a71662b03bf1e75e7637a24da41b919
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1641545
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#665939}


CrOS-Libchrome-Original-Commit: f808a5b405f1c8674ddaf89bc8e715c554080c9a