[scheduler] Adding base::CurrentThread trait.
I'm adding a trait specifying that the task should run on the current
thread. It will be one of the 3 possible destinations (alongside
ThreadPool and BrowserThread), specifying which will be made mandatory
in a future CL.

A future CL will add the desired functionality backing this trait.

Bug: 968047
Change-Id: I16f244911b9e25da354148b034bd766c988dec11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642559
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Karolina Soltys <ksolt@chromium.org>
Reviewed-by: Alex Clarke <alexclarke@chromium.org>
Commit-Queue: Karolina Soltys <ksolt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666288}


CrOS-Libchrome-Original-Commit: 973df87d674a9453e36028f3801e67306dc94203