Switch policies to SCHEMA_ALLOW_UNKNOWN
This CL switches schema validation for all policies that use a
SchemaValidatingPolicyHandler in combination with SCHEMA_STRICT to
SCHEMA_ALLOW_UNKNOWN. The resulting change will be that policies with
an unknown field will still be accepted instead of being rejected. The
unknown field will be removed when the policy is applied (by
Schema::Normalize() in
SchemaValidatingPolicyHandler:CheckAndGetValue()). The unknown field
will still be visible in the value displayed on chrome://policy and
will also generate an error message there.
This CL also changes validation/application of component policies from
Validate() with SCHEMA_STRICT to Normalize() with SCHEMA_ALLOW_UNKNOWN.
This means that policies with unknown fields would not be rejected
anymore, the unknown fields are removed from the applied policy though.

Reason for this change is that SCHEMA_STRICT would not allow the
policies to change in the future by adding a new field.

Bug: 953615
Change-Id: Ide2f25ec5fa2f964110406c8f94d0e811eda0c2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1640553
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: A Olsen <olsen@chromium.org>
Commit-Queue: Alexander Hendrich <hendrich@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671069}


CrOS-Libchrome-Original-Commit: 9fce8ee92bdb4b1d6a2ebba68b88f296fe3e6141