Add a ScopedRunTimeoutForTest on the main thread in BrowserTestBase
This causes browser tests to encounter a gtest failure, including
helpful diagnostics such as the call-stack, if they become stuck in a
RunLoop::Run() call, rather than relying on the TestLauncher killing the
test batch process (which only generates a stack trace on some
platforms).

This CL also fixes a bug in the content test_launcher which was
incorrectly using action_max_timeout() to launch tests (it should use
test_launcher_timeout()). This bug dates back to 2011 when this code was
first introduced in crrev.com/bb36d82adbfa2 (so it looks like an
oversight rather than an intentional usage of action_max_timeout()).

This CL also corrects action_max_timeout()'s documentation. It was
incorrectly stating that this timeout could be waited upon multiple
times in a test (but it was 30s and test_launcher_timeout() is 45s).
Luckily no test did wait multiple times for it in a test (at least not
without generating a failure after the first timeout).

And adds tests that the RunLoop timeout is always less-than
test_launcher_timeout() -- catching a bug that it wasn't on
ios_simulator under !NDEBUG. Fixed by removing unnecessary NDEBUG
exception for action_max_timeout().

Bug: 918724, 922098
Change-Id: Id4cfdafa99f0e1d5080181a0831a80cd1d2e98da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1616781
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672973}


CrOS-Libchrome-Original-Commit: c38fc01b71723cb0cf889edc9b147dabf0ca7115