Make base::SequenceBound move constructor parameters from their storage type.
Previously it would forward their original parameter type, which might
be (for instance) an lvalue reference. Instead we should call the
constructor using the types that we have available in the BindState
storage, which is std::decay_t<Args>&& (since the invoker for BindOnce
moves from the storage).

As reported in #cxx on Slack.

Change-Id: If7afb9722ba1fa3d19913915ac34b582c5908d23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1682320
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674210}


CrOS-Libchrome-Original-Commit: c6d87d9ad8ba955dab3c828a05cf82f1294e557c