crazylinker: fix java tests and add a doc
I found that crazylinker tests are special enough not to be run during
regular testing. I fixed a small issue and added a document about a few
non-trivial-to-discover ways to test the crazylinker in Chromium
checkout.

After the fix all the off-CQ testing described in the doc passes on my
Nexus5X, with the exception of fuzzers - I did not run those.

Applying TBR for docs only.

TBR=rmcilroy@chromium.org

Bug: 981599
Change-Id: I7ead0ef59c6b402729393ba273798107afefed46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695309
Reviewed-by: Egor Pasko <pasko@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676528}


CrOS-Libchrome-Original-Commit: f2f3e06f771c35acc70bd588ac31eaf121eaad54