Move named shared memory to ServiceProcessState.
The only use for named shared memory has been for
ServiceProcessState to globally record the existence of a service
process. This CL removes named shared memory from the legacy
shared memory API, and puts it in the platform-specific parts of
ServiceProcessState, and updates that to use the new shared
memory region API.

This causes no functional change; shared memory continues to be
used to record service process existence on the same
platforms. The named shared memory is just accessed directly and
not through a SHM API. This CL does make it explicit that the
named region can be opened writable; this was true before but
obscured by the legacy shared memory API.

Bug: 396390, 795291
Change-Id: Iab6f6a0cbc7d85997e7ee536406af4ef1561c758
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1655437
Commit-Queue: Matthew Cary (CET) <mattcary@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677700}


CrOS-Libchrome-Original-Commit: 1a9636ac3e75db26abf7187e45ceb9a7f0f67c26