[base] CROS: MemoryPressureMonitor can run on shutdown.
This CL adds a more correct fix for https://crbug.com/964569 ,
specifically there is nothing wrong with the task remaining running
on shutdown. As the process exists this will break out of poll(2)
and the thread will exit. The only other option would be to try to
do an overly complicated shutdown dance which is completely unnecessary.


BUG=chromium:964569
TEST=Validated on a betty vm.

Change-Id: I0f32b72f8c1db595f0b0d9cce7fd06759430a2f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1721409
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Brian Geffon <bgeffon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681523}


CrOS-Libchrome-Original-Commit: 746f7a3f2f23b2d3f7407093c0379fda4feb4746