[TraceLog] Fix uninitialized members
Also renamed the types to "Functions" instead of "Callbacks" as the
latter is confusing with a properly bound base::Callback where the
former clearly denotates a function pointer.

This lack of initialization resulted in an access violation instead of
using the non-overridden function when tracing is called in contexts
where the perfetto override isn't in place yet as part of
https://chromium-review.googlesource.com/c/chromium/src/+/1680893

We were simply lucky to get away with until now.

R=eseckler@chromium.org

Bug: 899897
Change-Id: I815fe5197888c32200cdf5166a0be4fbf79d355a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1721071
Auto-Submit: Gabriel Charette <gab@chromium.org>
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681769}


CrOS-Libchrome-Original-Commit: e1d1b3c918bcc35cf4feccad83de4e54175b6275