[ThreadPool]: Create JobDelegate interface.
This CL exposes new interfaces publicly (under the experimental namespace) for the Jobs API.
JobTask: Base interface to implement a job
JobDelegate: entry point that's passed to a worker task to communicate with
  the scheduler.

The implementation of JobDelegate is partial. The real implementation,
as well as usage assertion is in:
https://chromium-review.googlesource.com/c/chromium/src/+/1713146

This CL also refactors MockJobTaskSource and related tests to encompass
the new types, and to respect the incoming assertion that max concurrency
decreases unless NotifyConcurrencyIncrease() is called.

Bug: 839091
Change-Id: I34adeeb046a2aca28288c82dc382ef1cb6717b84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707100
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681921}


CrOS-Libchrome-Original-Commit: c6b18ebca49c339016272661cefd0683a92fef36