Create stub for NativeUnwinderAndroid
This class will contain the Android-specific unwinding implementation,
which will rely on libunwindstack. libunwindstack will be pulled in as a
dynamic feature module (DFM).

Having a stub version of this class will allow us to create
StackSamplerAndroid. (StackSamplerAndroid will need to create a
StackSamplerImpl, which accepts an Unwinder as a constructor parameter.)

TBR=gab@chromium.org
R=wittman@chromium.org

Bug: 988574
Change-Id: Id929731e94232f457add7914eb426ebcdb0bf0f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1726829
Reviewed-by: Charlie Andrews <charliea@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Mike Wittman <wittman@chromium.org>
Commit-Queue: Charlie Andrews <charliea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682404}


CrOS-Libchrome-Original-Commit: 98447da8f9ab0998ac68d42133725c302d52e68a