Add --disable-highres-timer command line switch.
The new command line switch allows users to disable high frequency
timers on Windows. Normally the browser will set a system-wide timer
resolution of 1ms on AC power, and 4ms on battery. The period used when
on battery power can be increased to 8ms if the experimental
SlowDCTimerInterrupts feature is enabled.

The switch is intended for use with headless browser processes running
on a server or VM, which is likely to be on AC power at all times. The
switch allows these processes to opt-in to the slower timer interrupt
frequency, acting as if they are running on battery. This should reduce
the amount of context switching and reduce CPU usage, which can be
helpful when attempting to run many headless browser processes in
parallel.

Bug: 986338
Change-Id: I8580f486aca3400fb51e0c4932c3783332a3a11a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1715566
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Brandon Walderman <brwalder@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#684119}


CrOS-Libchrome-Original-Commit: ac7514b02c7b01fe56a60f75387b63aa4f00a66e