Add switch-dependent feature overrides to FeatureList initialization
This adds the ability to set up feature overrides that depend on the
presence of a command-line switch. These overrides are processed
during FeatureList initialization, after explicit enables/disables,
but before field trial setup. This allows the initialization of
a base::Feature according to a command-line switch, which is superseded
by explicit --enable-features and --disable-features, but which
supersedes variations/field trials.

Bug: 988603
Change-Id: Id808325d94689d272b9bc00b400b44d0560d6136
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1724704
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685668}


CrOS-Libchrome-Original-Commit: d49e375109b214ef1e66fae7abf84157e0931e83