Add operator*() definitions for mojo::{Associated}Remote
Having this operator defined would enable the possibility of creating
bound functions off a moved remote, similar to what can already be done
for InterfacePtr<T> types, but for mojo::{Associated}Remote types.

For instance, the following code (that exists in the current codebase):

  InterfacePtr<T> ptr;
  Bind(&T::Foo, std::move(ptr));

... could be migrated into the following snippet with this new definition:

  mojo::Remote<T> remote;
  Bind(&T::Foo, std::move(remote));

Note that this is not something that you should normally be doing as most
times a PendingRemote<T> should be moved instead of a Remote<T>, but this
is still useful in some cases like testing, where a Remote<T> must be
constructed off a PendingRemote<T> (and therefore become bound) first, and
then construct a bound function out of it that will be used later on.

You can see see MockStorageArea::GetAll() in mock_storage_area.cc [1] for
an example of this kind of situation, which motivated this CL.

[1]https://cs.chromium.org/chromium/src/third_party/blink/renderer/modules/storage/testing/mock_storage_area.cc?l=68

Bug: 978694
Change-Id: I79b748b8f67d6a2c179eebf0875d83c6816dce88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1742092
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Cr-Commit-Position: refs/heads/master@{#685948}


CrOS-Libchrome-Original-Commit: adca65e3f933d55823689e03fa7c2711099c01de