Use StringPiece for the optional suffix in base::GetUniquePathNumber.
This removes the need to create/destroy a string when calling. This
change also adds a default value of an empty StringPiece to the suffix
so that callers who don't care about a suffix don't need to bother with
it. Finally, this change adds a unit test for the function, which it
apparently never had.

BUG=none
TBR=vasilii@chromium.org,vakh@chromium.org,eladalon@chromium.org,jamiewalch@chromium.org,carlosk@chromium.org

Change-Id: Iaf98d5237ffcc211bf9dfc2242742508f35bafac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1752302
Commit-Queue: Greg Thompson <grt@chromium.org>
Auto-Submit: Greg Thompson <grt@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686754}


CrOS-Libchrome-Original-Commit: 42373544a5f808d4c68e6588c1eb1604e02eb8db