[Fuchsia] Implement sysmem buffer collection import for GPU process.
FuchsiaVideoDecoder (which runs in the renderer) will need to share a
sysmem buffer collection with the GPU process in order to render
decoded video frames. This CL adds IPC messages for the renderer to
use in order to import sysmem collections to the GPU process. It also
updates VulkanImplementationScenic to make it possible to create Vulkan
images from the imported sysmem buffer collections.

Bug: 981026
Change-Id: Id1ae9b9dc521feddd6eb2e41847b66617384062b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745716
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687190}


CrOS-Libchrome-Original-Commit: 7b343ffec9df69278f4bd3a5476dc24921b7a811