[TaskEnvironment] Execute rename plan for ScopedTaskEnvironment
ScopedTaskEnvironment => TaskEnvironment

SingleThreadTaskEnvironment ==
    TaskEnvironment{ThreadingMode::MAIN_THREAD_ONLY}

TestBrowserThreadBundle => BrowserTaskEnvironment

Updated documentation.

The filenames were left as-is in this file and will be renamed in a
follow-up CL that mass renames users on top-of-tree.
We could have moved the files and added back scoped_task_environment.h
for the redeclaration but that would have clobbered git history. So
this approach was preferred.

Lastly, had to redeclare old names as subclasses instead of using
using-decls because using-decls aren't compatible with fwd-decls used
in a few tests throughout the codebase.

R=dcheng@chromium.org

Bug: 992483
Change-Id: Id408970df116ca66a1c2e150f7d2683aa93cac2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1747171
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687238}


CrOS-Libchrome-Original-Commit: 90a5675cbfd57a413ff727d94d711856c31dfab2