GetUniquePath{,Number} cleanups.
GetUniquePathNumber previously took an optional suffix that was poorly
documented and confusing to use. This change modifies the one caller
that used this suffix (in chrome/browser/downgrade) so that it is no
longer needed and removes it. The documentation for GetUniquePathNumber
has also been revised so that it is now accurate and comprehensible.

GetUniquePath previously returned the original path in case a unique
name could not be found. This was surely a bug, and is now fixed. The
documentation for this function has also been revised for clarity and
accuracy.

Furthermore, the callers of GetUniquePathNumber that did nothing more
with the result than format it into a new FilePath have been converted
to using GetUniquePath, thereby removing some duplication and
simplifying code.

R=eladalon@chromium.org

BUG: 785333
Change-Id: I7d75e615f58c1bc5f95b1edb7a744a3fc389df61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761596
Auto-Submit: Greg Thompson <grt@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Elad Alon <eladalon@chromium.org>
Commit-Queue: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689770}


CrOS-Libchrome-Original-Commit: b4abcb4ef5bec039830d15446b0cbbeb8b816165