Clarify how nullable types map in C++
Presently, the handling for an IDL-nullable type "Foo?" is buried
within the type mapping documentation for nullable_is_same_type,
making it easy to overlook.

Update the informational table that serves as a quick reference for
IDL to C++ types to include a note about it, as well as a reference
to the customizability.

Bug: 998985
Change-Id: If4e6609da44d1c2cbb81c32a1712d032e2c3a13a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1774330
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#691670}


CrOS-Libchrome-Original-Commit: 9353d6dcefe1de9a73d852f6f3e9bd3163f25b48