Android: Ensure that no tests run during instrumentation test listing.
To support parameterized tests, our test runner does not use the default
test listing logic, and instead requests that tests be run, and relies
on BaseJUnit4ClassRunner to no-op them all when in this mode.

This scheme works only if tests actually use BaseJUnit4ClassRunner (or a
subclass). The consequence of not using this runner is that the tests
will run during test listing.

This changes our test runner to enforce that all tests use
BaseJUnit4ClassRunner so that we never have tests mistakenly being run
during test listing.

Bug: 999007
Change-Id: I226fe55b5716c713d47a6027c088ca6774ae0c10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1777089
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692128}


CrOS-Libchrome-Original-Commit: 63ba14a130f3ef3a99d0d275f8c69243cc92bdc9