android: Load the trichrome's native library with ModernLinker.
Trichrome doesn't benefit from the automatic relocation sharing that we get from
the WebView path for Monochrome. As a consequence, we don't have relocation
sharing on Android Q, with a memory cost of 2MB per process on ARM32, and more
on ARM64.

This uses android_dlopen_ext() to share relocations for Trichrome, by:
- Adding the chromium linker to the trichrome (and monochrome) builds
- Using it for Trichrome

This change also rolls back the use of ModernLinker for Chrome{,Modern} on N+
that was added recently to ease local and waterfall testing (see attached
bug). From this commit on, ModernLinker is only used for Trichrome.

This forces us to change the library loader code, as JNI registration is lazy
for Monochrome, that is performed by ART just in time when a new native method
is executed. ART essentially walks the list of all lib handles it knows about,
and tries to find the matching native symbol with dlsym(). This means that for
Monochrome, JNI symbols are exported, but manual registration is compiled out.

However, ART doesn't have a handle on a library opened with android_dlopen_ext()
by Chrome, and cannot find symbols this way as a consequence. There are then two
choices:
1. Perform manual JNI registration
2. Make ART aware of the library

The first choice would lead to a ~135kiB size increase, as the symbols names are
present twice (since we keep lazy registration for Monochrome and Monochrome and
Trichrome have the same library). The second approach is what we do, by loading
the library with System.loadLibrary() after the android_dlopen_ext(). This makes
sure that ART has a handle on the library, while avoiding a double load or a
size penalty.

This commit still increases the library size from adding the chromium linker to
{Mono,Tri}Chrome. This penalty will be significantly reduced in a forthcoming
commit by stripping out crazylinker from these builds, which is the bulk of the
chromium linker, and that we don't use here.

Finally, ModernLinker still writes relocations to disk. Using memfd_create() to
avoid this is planned for privileged processes.

Change-Id: Iccb936cb4c6db51654679e312bb08bf9c2846101
Bug: 979638
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1733553
Reviewed-by: Benoit L <lizeb@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693254}


CrOS-Libchrome-Original-Commit: 35c0e467054f747c7dcd3a9d651e4eb52ccd9dbf