[base] Replace GetList().{emplace,push}_back with Append in //components
This change replaces usages of base::Value::GetList() followed by either
emplace_back() or push_back() with base::Value::Append(). This is
because of the upcoming change to GetList() to return a base::span
instead, which does not support either emplace_back() or push_back().

This is a completely mechanical change. Steps to reproduce:
- sed -i 's/GetList().push_back/Append/g'
- sed -i 's/GetList().emplace_back/Append/g'
- git cl format

Bug: 646113
Change-Id: Ia29bc02040a056e518737f2b11f1cd0b66b0e0f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1796430
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695530}


CrOS-Libchrome-Original-Commit: d322b291d93d5220b02cb16643fcd7ca62c30fec