[base] Replace GetList().{emplace,push}_back with Append
This change replaces usages of base::Value::GetList() followed by either
emplace_back() or push_back() with base::Value::Append(). This is
because of the upcoming change to GetList() to return a base::span
instead, which does not support either emplace_back() or push_back().

This is a mostly mechanical change. Steps to reproduce:
- sed -i 's/GetList().push_back/Append/g'
- sed -i 's/GetList().emplace_back/Append/g'
- git cl format
- Manually fixing the breakage in base/json/json_writer_unittest.cc
  and tools/json_schema_compiler/util.cc

TBR=afakhry,boliu,eugenebut,mmoroz,rdevlin.cronin,rockot@google.com,sandersd

Bug: 646113
Change-Id: I2f13fd0efa7fe712f7659500404192be4a344e14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1796427
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695711}


CrOS-Libchrome-Original-Commit: 85a66712acce954820b07e4c43b1e8f6d0accdde