[base] Rename CheckedRandomAccessIterator to CheckedContiguousIterator
This change renames base::CheckedRandomAccess(Const)Iterator to
base::CheckedContiguous(Const)Iterator.

This change is motivated by the fact that this iterator not only
fulfils the requirements of RandomAccessIterator [1], but also
ContiguousIterator [2].

Furthermore, the implementation of the iterator assumes it is pointing
into a single contiguous array, and thus is unsuitable for containers
that have random access, but don't operate on an underlying array, such
as std::deque.

[1] https://en.cppreference.com/w/cpp/named_req/RandomAccessIterator
[2] https://en.cppreference.com/w/cpp/named_req/ContiguousIterator

Bug: 817982
Change-Id: I7186dfc3b2849818b9f83d62e05ebc5a79aaf5ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1811377
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Chris Palmer <palmer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697793}


CrOS-Libchrome-Original-Commit: 27147e685c9dbe705158a1ae7540c28598bae9b2