[android] Init ICU with extra data if extra ICU module is installed
The purpose of this CL is to provide ICU support for 30 extra languages
(see crrev/c/1648623 for more details). crrev/c/1811973 adds a feature
module that delivers an extra ICU data file containing support for the
aforementioned extra languages. To make use of this extra data file, we
check at browser process startup whether the extra ICU feature module is
installed and if so init ICU with the extra data file inside the module.
We retrieve the module install state from the ContentBrowserClient as
feature modules are a chrome layer concept and the ICU initialization
happens in the content layer. The extra ICU data file is applied
_before_ the main ICU data file. Otherwise, the main ICU data file would
override the extra languages.

Requires crrev/c/1834464.

Bug: 1006794
Change-Id: Ic7c14bf3437309b90cf2ce205ade14da937203c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1818876
Commit-Queue: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703491}


CrOS-Libchrome-Original-Commit: bf1b96ef042cce050184bca35c35dafcd3f91784