Moved switch definition from unit_test_launcher.cc to test_switches.h
Moved the "single-process-tests" switch definition from the
unit_test_launcher.cc to base/test/test_switches.h and
base/test/test_switches.cc to conform to the convention used for other
test flags in the base project.

The motivation for this change to be able to reuse the
"single-process-tests" switch outside the base/ project.

Bug: 460513
Change-Id: If8fe29a790b51255ec128a93484e4cd494de373a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1845959
Reviewed-by: kylechar <kylechar@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703823}


CrOS-Libchrome-Original-Commit: 41ffef1ba7181f4fa170210b113f2f0365fe0b9e