Add base::Value::AsDictionaryValue().
This is an adaptor to allow a caller to use base::Value instead of
base::DictionaryValue to construct a dictionary value, and able to
easily pass the value to a function that still expects const
base::DictionaryValue&. This makes it easier to move away from
base::DictionaryValue in smaller increments. Use it in one place to
demonstrate it works.

Do the same for list values with base::Value::AsListValue().

Bug: 646113
Change-Id: I9116525ef5a5a80d3752cdca1960d21b2288ff42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1838684
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703845}


CrOS-Libchrome-Original-Commit: 8c1432bfda40eeffaef524d35ca4749afeb9d24d