Log printer information in PrintJobWorker::UpdatePrintSettings().
PrintJobWorker::UpdatePrintSettings() is used for printing, but does not
log any information in case of a crash. Add a ScopedPrinterInfo here to
capture printer driver info for the duration of the call, so crashes in
this code path can give more clues as to what is causing the crashes.

Bug: 1013433
Change-Id: Ic6e1dab3cb40a70c408fb8a0a3c6d08601b24cb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1855298
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705142}


CrOS-Libchrome-Original-Commit: feb597bb999acaf4ce05ca4ae1930cc939c6638d