Add scoped_refptr(nullptr_t) constructor.
Add a new scoped_refptr constructor for creating from nullptr. This is
to better understand how many callers rely on the scoped_refptr(T*)
constructor and what would need to be fixed for it to be made explicit.

Also fix a few places where NULL is still being implicitly converted to
scoped_refptr<T>. This causes a compilation error with the new
constructor as NULL is defined as 0.

TBR=marq@chromium.org

Bug: 589048
Change-Id: I117703b4db3472e191bf64629ec7ac7966ba10d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1876908
Commit-Queue: kylechar <kylechar@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Sean Topping <seantopping@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711100}


CrOS-Libchrome-Original-Commit: de57aa352172a46b99c1c6e9c924fe81e751f437