[Android] Introduce AppMenuTest and AppMenuAdapterRenderTest
Adds an AppMenuTest for app menu integration tests
 - #testShowHideAppMenu
 - #testShowDestroyAppMenu
 - #testClickMenuItem
 - #testAppMenuBlockers
 - #testSetMenuHighlight_StandardItem
 - #testSetMenuHighlight_Icon
 - #testMenuItemContentChnaged
 - #testHeaderFooter
 - #testAppMenuHiddenOnStopWithNative
 - #testAppMenuHiddenOnConfigurationChange

Also adds a few more tests to AppMenuAdapterTest to shore up test
coverage and adds an AppMenuAdapterRenderTest that runs a suite of
render tests for the various types of menu items supported natively
by the app menu.

This patch includes an update to AnimationFrameTimeHistogram to allow
histogram recording to be disabled for tests.

BUG=966644

Change-Id: I035b10557a9e35ef0e86b315ccd926230bb18bf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1893984
Commit-Queue: Theresa  <twellington@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712702}


CrOS-Libchrome-Original-Commit: a311f5a36af57c83b5e4fc9702ac36905d6d79f9