Modules: Always resolve native library paths through ClassLoader
Normally, the system linker and Crazy Linker can resolve libraries.
However, if a module is installed via SplitCompat, and hasn't yet been
fully installed, then the module (and libraries) reside at an alternate
location unknown to the linkers.

In order to find these libraries, we need to call on the Java
ClassLoader. This is already being done for production libraries, but
not in the component build case (currently affecting only the Test
Dummy module).

As a result, bundle smoke tests running with "fake install" would fail
with component builds - that scenario is meant to look like the
SplitCompat'ed case.

Bug: 1025316
Change-Id: If2bcf2de85b323e5e14ed70561e3f31d50bf8287
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1922276
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716475}


CrOS-Libchrome-Original-Commit: cd6ec21e5a5c759cf3f31b6cc15ce4d682ef59ef