[sequencemanager] Track low priority and blocked tasks in TaskObserver.
This CL adds a boolean argument to base::TaskObserver::WillRunTask(),
to indicate whether the task that is about to run was at any point
in a queue that was disabled, blocked by a fence or low priority.
The purpose of this change is to allow the creation of an histogram
that reports when tasks that were never blocked or low priority are
queued for a long time.

TBR=sky@chromium.org,avi@chromium.org,zmo@chromium.org,marq@chromium.org,mattm@chromium.org,jochen@chromium.org

Bug: 1029137
Change-Id: If63e2a5cd5690e04663399643ba2a3f7d928cedc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1943538
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Alex Clarke <alexclarke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721611}


CrOS-Libchrome-Original-Commit: c0e03043c0304e127506db04d6fd6c5073923281