chromecast: Migrate IS_CHROMECAST to common buildflag
The IS_CHROMECAST define is declared in 16 different places, which is
risky especially if the define is used in a header file.

Move this to a common buildflag header in //build, alongside the chrome
branding flags. This flag needs to be in a widely scoped location, since
we don't know if we're allowed to use //chromecast until we check the
flag's value.

OS_CHROMEOS uses a global #define for this, and that would work here, but
global ifdefs are discouraged. There are few enough uses of IS_CHROMECAST
that having explicit dependencies on a buildflag header isn't too onerous.

No functional change.

Bug: 1030969
TBR=kylechar@chromium.org, dgozman@chromium.org, benwells@chromium.org, kbr@chromium.org, dalecurtis@chromium.org, reillyg@chromium.org, foolip@chromium.org

Change-Id: I48d9ffccde7e7a680ce572f15a8ba1ef11adcc6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1951981
Commit-Queue: Michael Spang <spang@chromium.org>
Reviewed-by: Luke Halliwell (slow) <halliwell@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722665}


CrOS-Libchrome-Original-Commit: 8a063346f10774b61fb54eb69e47a2fa018193b1