android: Share relocations with shared memory in ModernLinker.
ModernLinker currently uses a regular file to share relocations. This is due to
its use of android_dlopen_ext(), which cannot use an ashmem-backed file, for the
following reasons:
- On the writing side (creating relocations), it write()s to a file, whereas
  ashmem regions have to be pre-sized.
- On the reading side (applying existing relocations), it calls mmap() with
  MAP_PRIVATE, which doesn't work with ashmem.

memfd_create() is a solution, however memfd-backed fds cannot be passed to
isolated processes as of Android Q, due to a SELinux restriction for isolated
processes.

This commit avoid unnecessary disk writes by using an ashmem-backed fd "after
the fact". More precisely, on the browser side:
1. Load the library at a fixed address using android_dlopen_ext(). Do not ask it
   to create shared relocations.
2. Find the relocations from the program headers
3. Copy them to a new ashmem-backed file mapping
4. Use mremap() to replace the current relocations with the newly-created
   ones. Note that this and the previous step work, as relocations are
   read-only, and mremap() is atomic.

On the child process side:
1-2. As in the browser case.
3. Map the relocations, and compare them to the current one.
4. If they match, use mremap() to replace the current relocations with the new
   ones.

This is not quite as good as what android_dlopen_ext() does, as there is no
support for partially-matching relocations. This will be added later if required.

From a practical standpoint, this CL:
- Adds calls to libandroid.so for the shared memory functions.
- Removes file-related code
- Expands program headers parsing
- Adds the mechanism mentioned above

Bug: 979638
Change-Id: I494b59fae453b4cfa93ccba9dbb3f8e72482c802
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1771599
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722919}


CrOS-Libchrome-Original-Commit: aedec224d0a798bdb7da307041f149a1acbd23c6