Android: Reset PathUtils statics in Robolectric tests
Robolectric runs each test with a different dataDir, so we need to be
resetting the cached directories in PathUtils between tests.

Noticed this specifically in WebappDirectoryManagerTest.java, where the
test was having to manually wipe the directory in order to pass.

Bug: None
Change-Id: I34944bd01ad39991593e3021c22e13dac3279f19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1956051
Commit-Queue: Peter Wen <wnwen@chromium.org>
Auto-Submit: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723100}


CrOS-Libchrome-Original-Commit: 185d9352a972a5dfce79f82b58ff0840def806e2