Android: Change StackTrace::ToString() formatting to match OS
StackTrace::ToString() prints its output in a format different from that
of Android's debuggerd crash stack output. That means tools that decode
stacks (eg. stack.py) need to handle multiple formats.

Android's trace format is reasonably easy to decode, even if we don't
know exactly which APK was loaded.  StackTrace:: is missing some of that
information (specifically the offset of the library within the APK), and
hence, if libraries are loaded directly from the APK, we've lost
fidelity.

This change brings StackTrace::'s output format in line with what
Android dumps, in hopes that decoding tools (in general) can be simpler
and more reliable.

Bug: 986622
Change-Id: Ib3c09a5a97eec6e5dd2c0989e609da152e05f30b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1953698
Reviewed-by: Christopher Grant <cjgrant@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Christopher Grant <cjgrant@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723112}


CrOS-Libchrome-Original-Commit: 679bc6479d78041041b98dab57fdb6dd78fb1c92