Prevent SysInfo::GetHardwareInfo from blocking shutdown
This changes the traits of the task posted on Windows in
SysInfo::GetHardwareInfo to make sure that this task doesn't block
shutdown. On Windows this task relies on WMI which can take a really
long time to initialize (several minutes in some situations) and so
it's better to use the CONTINUE_ON_SHUTDOWN traits.

Bug: 1035617
Change-Id: Iaba73c20d3e20a1affbb05bb89b5cd08e2618f52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1988223
Auto-Submit: Sébastien Marchand <sebmarchand@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729164}


CrOS-Libchrome-Original-Commit: 0b45105d4d098c029226ea7b9f29efc74d3117ff