Android: Better error messaging for some UnsatisfiedLinkErrors.
A common cause of UnsatisfiedLinkErrors is that the native library
hasn't been loaded yet. It's difficult to tell the difference between
this and other UnsatisfiedLinkErrors.

This CL adds generated code that queries LibraryLoader to determine
library loading status and throws an explicit error when a native
method is called before the library has been loaded.

Bug: 1022938
Change-Id: I63980990abaef16da3847a05eda01cb116779560
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1976905
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Eric Stevenson <estevenson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#729461}


CrOS-Libchrome-Original-Commit: 24f7b916a7aee9a42682f5f39d331773b8f56e69