Mojo: pass arrays by value to setters and wrapper constructors
This changes the type for std::vector arguments passed in wrapper class
constructors and setters from reference to value (i.e. from
const std::vector<>& in_param to std::vector<> in_param), so that

std::move() that we invoke in the bodies of constructors and setters
are actually effective.

The effect of the change on generated files can be seen here:
https://paste.googleplex.com/5618555812839424?raw

Change-Id: Id9cfd3bf23257d2418d462f52a39f3b01e1de96c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1990446
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730067}


CrOS-Libchrome-Original-Commit: f32f002b95fbd31afc1fde47d0b2a46c9fc3360e