Reformat all gn files in /mojo.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=oksamyt@chromium.org

Bug: 1041419
Change-Id: I23ed7710dad51f52eb56d7369da9feb9b2275643
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998242
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731306}


CrOS-Libchrome-Original-Commit: 11fc53eb77b11ab43437c15d36bfdcd2b3ea73b1