[i18n] Adds the setup for tzdata loading
The tzdata is loaded only for OS_FUCHSIA, since Fuchsia is the only OS
that needs this.  Fuchsia needs tzdata loading because it supports
multiple runners, and it is important that all of them use exactly the
same tzdata version.

If the tzdata files are not present, the execution proceeds, but the
files are not loaded.

In production, the runner will get time zone data from the config data
directory which is made available to apps in specific Fuchsia products.

Two unit tests are added:
1. A unit test that confirms tzdata loading affects the reported library
  revision.

  This test confirms that the loading mechanism works.

2. A unit test that checks whether the loaded timezone data revision
  matches what Fuchsia expects to see.

  This test will fail if loading fails so that the version in use is
  different from what Fuchsia expects.  This allows Fuchsia engs to
  see how to fix the harmful version skew.  This falilure should happen
  only if ICU libraries in Fuchsia and Chromium become backward-incompatible,
  which is unlikely but possible.


R=dpranke@chromium.org, wez@chromium.org

Bug: fuchsia:37487
Change-Id: I884777befbdb0ea9e324acc6c4359145c54d02aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949120
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731684}


CrOS-Libchrome-Original-Commit: bfdd3068d143714e032aa326e76f03abf152b9df