Remove DCHECK causing test flakiness
The resident size recorded from the metrics show it is never 0. The
devcies in the lab might get a 0 value if some code tries to override
mallinfo. So, remove this unnecessary check. If the metrics go wrong in
future, we can detect that from the UMA.

BUG=843843

Change-Id: Ie9a2e9f5e3f18b1572e45ef8db291f155dbe7c7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2004903
Auto-Submit: ssid <ssid@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#732603}


CrOS-Libchrome-Original-Commit: 4e3e77b937660d6a504c62586ff0118970b0b3bb