Fix Py3 compat in vs_toolchain.py
In Python 3, dict.keys() returns a generator instead of a list. Call
list() to turn it into a list so that it can be indexed.

I bravely decided to put Python 3 on PATH, and I needed this in order to
re-run gn gen. This seems to be the only change needed to run `gn gen`.
Later on in the build there are Unicode errors during mojo binding
actions.

R=brucedawson@chromium.org, hans@chromium.org

Bug: 941669
Change-Id: Ib80e11798003999469776122b0a32f416909565b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2013514
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Reid Kleckner <rnk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733830}


CrOS-Libchrome-Original-Commit: 8653c97f6c7a4c268682a359c4ad3523d426eb64