Add SourceIdType::WEBAPK_ID for recording webapk ukm
Whe recording WebAPK UKM, we use manifest url as the source_url. But UKM
has a mechanism that filter out the url that are not whitelisted.
Manifest url does not matches the navigation url or other whitelisted
url, so the record is drop with DropDataReason::NOT_MATCHED.

This CL adds another SourceIdType WEBAPK_ID for the WebAPK manifests, so
that these url are whitelisted.

Bug: 1044706
Change-Id: I988473f52afcf5b25cc5fafd43934e866f628f76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2018086
Commit-Queue: Ella Ge <eirage@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Reviewed-by: Glenn Hartmann <hartmanng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736903}


CrOS-Libchrome-Original-Commit: ed2f42c2c18c3237c0e2807256a635a445404d8a