[base] Minor CheckedContiguousRange improvements
This change improves the documentation for CheckedContiguousRange and
makes its default constructor well-behaved. Prior to this change
CheckedContiguousRange was default-constructible, but every operation on
it would have resulted in a nullptr dereference.

After this change a caller can query a default constructed range for its
data() and size() and obtain iterators without causing a crash.

Bug: 990059
Change-Id: I14e1a42c86591c4b362dd1e05ea21ed35c7610a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2049963
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#740316}


CrOS-Libchrome-Original-Commit: 3a2bcfd951651136c0bed66e58cf9d5ec16f4878