Android: Enable Jetify to move to androidx"
To update to androidx, we run the "jetify" script on all targets
passed to process_java_prebuilt. This converts all references in the
jar to androidx without having to modify the source code. This also
allows us to not have to check in third_party jars that reference
androidx, such as the gms libraries.

To compile with the androidx dependencies with the
android_support_library references still in the code, we modified the
java groups in android_deps/BUILD.gn to have the dependency target
depended on androidx, but use the old support library as an input_jar.

All the resources (such as layout.xml files) need to be converted to
androidx too, which was done in another CL, this CL just activates that
feature in compile_resources.py

Bug: 896775
Change-Id: Ib53c670b53145999ca57eebd4905e8f22684de67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1830506
Commit-Queue: benjamin joyce <bjoyce@google.com>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741505}


CrOS-Libchrome-Original-Commit: 76b28f21d8741df1bb1c9f4093fe5e633fd24bdc