Remove DisplayPrivate calls from ContextFactoryPrivate
Most of the functions in ContextFactoryPrivate are keyed on
ui::Compositor* to modify something for a specific display compositor.
The implementation forwards these calls to the corresponding
viz::mojom::DisplayPrivate. Instead, give Compositor a pointer to the
DisplayPrivate and make the calls directly.

Make VizProcessTransportFactory a ContextFactoryPrivate implementation
so that HostContextFactoryPrivate can be deleted.

For InProcessTransportFactory, move these calls into the
PerCompositorData object, which implements DisplayPrivate.

Bug: 1042259
Change-Id: I64f93e2024e2db74bd04da0ff4e4a2f7c82409d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2050892
Commit-Queue: Sean Gilhuly <sgilhuly@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Reviewed-by: Johannes Henkel <johannes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742293}


CrOS-Libchrome-Original-Commit: 55dbb2ceb82d305bb88663a5f20280064ac07c45