[base] Generalize and Simplify MoveArg
This change generalizes base's MoveArg by using a functor instead of a
GMock action. This allows it to have a default argument for the index,
and can also support more than 10 parameters. Furthermore, this change
removes other implementations of MoveArg in favor of the version in
//base.

TBR=danakj

Bug: 714018
Change-Id: I3a1fdf602d66481f301d4bd7049076b9d46718c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2066626
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743635}


CrOS-Libchrome-Original-Commit: 5d66f95cc1480b7f0b80319e8e5e9e08b7d68bf2