[Sampling profiler] ModuleCache interface cleanups
Consistently use 'const Module' rather than 'Module' in the interface.
Modules are not to be mutated by the ModuleCache or its users.

Rename InjectModuleForTesting to InjectNativeModuleForTesting to be
clearer about what it does.

Have GetModules() return both native and non-native
modules to make the behavior consistent with the function
name. The only current user of GetModules(),
content::protocol::MemoryHandler::GetBrowserSamplingProfile(), does
not add any non-native modules to the ModuleCache so returning
non-native modules from the function will not affect it.

Bug: 1035855
Change-Id: Ie1cd38e081ee48297df3c7a15b8e4997d1df1589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062883
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#744430}


CrOS-Libchrome-Original-Commit: 0f074d7b330f6a44fb1566001e631c95194896a2