[Android] Move IntentUtils.java to //base
This file is used from //chrome's ExternalNavigationHandler.java, which
we are componentizing for sharing with WebLayer. This CL moves
IntentUtils.java into //base/android, as it is conceptually
//base-level functionality.

This CL also removes this file from being listed in
chrome/android/java/static_library_dex_reference_workarounds.flags,
which prevents //chrome files from being used in Trichrome. That
restriction is no longer relevant now that this file is in //base
(especially since the goal of this work is *precisely* to end up
sharing this file with WebLayer ;).

Bug: 1031465
Change-Id: I7e4fd1ac42ba1083afbe0fc55582ee20b42d6596
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2071257
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745014}


CrOS-Libchrome-Original-Commit: be349452e47e472e5d072d3e8182c3e27dd53d05