android: Check classloader in bindToCaller
There is a situation where chromium java code is loaded into different
class loaders in the same process. The expectation is to avoid sharing
renderer processes between the two class loaders. So pass a string
representation of the class loader into BindToCaller and check it along
with the pid.

Bug: 1051358
Change-Id: I8147582026549e7a4749f63d1941bd47e7d67ce7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2051531
Commit-Queue: Bo <boliu@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#745540}


CrOS-Libchrome-Original-Commit: fe0f89f88a6bef5f547332e386ee86ba3ee7e145