Fix mojom parser tests
This fixes all the existing mojom parser tests that are still meaningful
and applicable and deletes the rest. This also fixes or (where
appropriate) disables any outstanding pylint issues in these flies.

The tests can be run locally with:

  python -m unittest discover \
      -s mojo/public/tools/mojom/ -p "*_unittest.py"

A follow-up CL will introduce a general-purpose script for registering
and running all Python-based tests related to Mojo tools so that they
can run on the CQ as well.

Bug: 1060473
Change-Id: Ic6d5083475089b8d49d9cbfc0ca55678d0f74928
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2098853
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#750026}


CrOS-Libchrome-Original-Commit: 77cb7c1e2a6250342e3b9349fd0d1f02eb676032