Enable Python-based Mojo unit tests on the CQ
This is a resurrected suite of Python unit tests covering the mojom IDL
parser. As efforts ramp up to improve, stabilize, and export the parser,
CI coverage of this code will become increasingly important.

A new GN target is introduced to aggregate the Python library and its
test dependencies, including the third-party libraries jinja2 and ply,
and a new mojo_python_unittests target aggregates all of this along
with a test runner script. The name is generic enough to encompass
tests for any future Python-based Mojo tools as well (e.g. for specific
generators in the Chromium tree).

The new test step (which currently runs about 90 tests in about 1.5s
on a cheap device) is added to a few desktop bot configurations.

Bug: 1060473
Change-Id: Ib38765acd148739e7a5edb94457a858d2490cf8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2100014
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750030}


CrOS-Libchrome-Original-Commit: 26a4046fe8f8afee7bf11cb55cdd5fb1e6711850