Relax PathService's existence requirement for directories.
PathService::Get was documented as guaranteeing that returned
directories existed. The getter for DIR_TASKBAR_PINS was recently
updated to check for existence and return false otherwise, but this
broke PathServiceTest.Get because it expected that paths can always be
retrieved. In this CL, we relax the existence guarantee in the API, stop
checking to ensure that the taskbar pins directory exists, and instruct
the test not to check for existence of this dir.

BUG=1053446

Change-Id: Ie0742bd191f6894494f5f2c7124b097dbb319eb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2102489
Commit-Queue: Greg Thompson <grt@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750338}


CrOS-Libchrome-Original-Commit: 92af8b99bdf5108e9eec35d1c42c111663c8f36c