Clean up locking in base/task
A lot of the locking in PostTask.java and TaskRunnerImpl.java was either
unnecessary, or was scoped much larger than it needed to be.

This change replaces most of the locks with either volatile or Atomics,
removing all locks around posting tasks, allowing multiple threads to
post tasks without waiting on each other.

Change-Id: Ic53d08325d859828e94b271129eea909af31ad00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096113
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750683}


CrOS-Libchrome-Original-Commit: b9cf574ed67324c6fe4af88066cac5cdb3460685