Fix base::JSONParser line and column count
Prior to this commit:
- The line count skipped line breaks within a JSON string.
- The column count was off-by-one for lines other than the first line.
- A missing ',' or '}' in a JSON object was also off-by-one.

For the existing
chrome/browser/chromeos/policy/device_policy_decoder_chromeos_unittest.cc
DevicePolicyDecoderChromeOSTest.DecodeJsonStringAndNormalizeJSONParseError
test, the invalid JSON data for that test is the "constexpr char
kInvalidJson" at the start of that file. That string is 13 bytes long:
https://source.chromium.org/chromium/chromium/src/+/master:chrome/browser/chromeos/policy/device_policy_decoder_chromeos_unittest.cc;l=13
The syntax error is the missing '}' (at what would be column 14), not
the final '"' (at column 13), so it seems more consistent after this
commit, where the JSONParser error's column is now also 14.

For the existing extensions/browser/sandboxed_unpacker_unittest.cc
SandboxedUnpackerTest.InvalidMessagesFile test, the invalid JSON data
for that test is
https://source.chromium.org/chromium/chromium/src/+/master:extensions/test/data/unpacker/invalid_messages_file/_locales/en_US/messages.json
which doesn't have a "line 2 column 10" that the test claimed, prior to
this commit.

For the existing extensions/common/file_util_unittest.cc
FileUtilTest.LoadExtensionGivesHelpfullErrorOnBadManifest test, the two
off-by-one bugs coincidentally cancelled each other out, so the test
code is unchanged. The invalid JSON data for that test is
https://source.chromium.org/chromium/chromium/src/+/master:extensions/test/data/file_util/bad_manifest/manifest.json

Bug: 1069271
Change-Id: I715185322c2a9a93964afe82944fa61928db155f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2154635
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761708}


CrOS-Libchrome-Original-Commit: 110c05ef73eccdf214f2342e87e422f5b7d1cd26