[threadpool] Report when a sequence holds more than 10000 tasks.
That suggests that there is code that posts tasks faster than they can
be executed by the pool, leading to unbounded memory usage growth.
These cases should be identified and fixed.

crbug.com/1071127 is a known instance of this problem.

PendingTask is 72 bytes. 10000 tasks ~ 720 KB

From the ThreadPool.NumTasksInSequenceOnPush histogram that was briefly
recorded in 84.0.4121.0, the max value recorded by a user in a day is
above 10000 tasks for less than 0.01% of users. Therefore, we don't
expect this change to flood the crash server.

Bug: 1071127
Change-Id: I08e4d44f36a816439931fc187a7753f3af4ddd1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151686
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#763872}


CrOS-Libchrome-Original-Commit: ce5b93b315c90eb3354d7b54720bd4ad106a305d