Move Messages for FirstRectForCharacterRange to frame.mojom
This CL moves TextInputClientMsg_FirstRectForCharacterRange to
LocalFrame and TextInputClientReplyMsg_GotFirstRectForRange to
TextInputHost.

Like [1], it adds GetFirstRectForRange to frame.mojom with
'is_mac' and replies with TextInputHost::GetFirstRectForRange
which runs on IO thread.

GetFirstRectForRange checks the caret bounds from the focused
plugin when it has ENABLE_PLUGINS. If it's disabled, it gets
the rect from the current frame selection.

[1] https://crrev.com/c/2015909

Bug: 1007365
Change-Id: Idb4b3b77a1c660679b1e09a46497ffd5ac462917
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111254
Commit-Queue: Julie Kim <jkim@igalia.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765412}


CrOS-Libchrome-Original-Commit: f43639291f32a8fc77f2e4382403280e983c1010