[base] Enable StringPiece conversions to std::string_view
This change allows implicit conversions from base::StringPiece to
std::string_view. As a result, base::StringPiece can be directly used in
multiple std::string APIs, without having to pass data() and size()
separately.

Note: This change works outside of C++17, because libc++ makes
<string_view> available since C++98. However, the NaCL toolchain does
not support this, which is why this conversion is not enabled on this
platform.

Bug: 1049498
Change-Id: I56c3a7a4bf165b67bc20502773ef13c33fddfadd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2189657
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766925}


CrOS-Libchrome-Original-Commit: 3314ea30abc476aced1e150e4f0b5076172a1e27