[AW] Add MetricsBridgeService for non-embedded webview processes
AwNonembeddedUmaRecorder can be used to collect metrics in
non-embedded WebView processes. It partially extends
base.metrics.UmaRecorder. In a follow-up CL it will be used as a
delegate recorder that will be passed to base.metrics.RecordHistogram
instead of being called directly.

MetricsBridgeService retains histogram records in memory as well as
saving them to a file for each read. It uses newly defined protobuf
messages for serializing and deserializing to and from disk as well as
over IPC calls.

This is hidden behind a feature flag "WebViewCollectNonEmbeddedMetrics"
that is disabled by default to avoid possible regressions in WebView.

Unit tests are added for Service, Recorder and replayer classes.

Bug: 1001185
Test: Using logs to verify that metrics are recorded and transmitted as expected
Test: out/Debug/bin/run_webview_instrumentation_test_apk
Test: out/Debug/bin/run_android_webview_junit_tests
Change-Id: I10da8c450f3585c267ba9dd77346586a21bebad0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2170888
Commit-Queue: Hazem Ashmawy <hazems@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#767486}


CrOS-Libchrome-Original-Commit: 3acf57a08ddd2100b08cedd61f40705c2077bfa6