[content] Add RFH/NH references to OnCookiesAccessed.
Add WebContentsObserver::OnCookiesAccessed(RenderFrameHost*, ...)
and WebContentsObserver::OnCookiesAccessed(NavigationHandle*, ...)
callbacks.

Attribute each cookie access to a particular RenderFrameHost or
NavigationRequest, which will allow us to attribute them to a particular
document.

This is needed because the network request made for a navigation can
read or write the cookie before the document is created and we have
to pass NavigationHandle* reference to avoid the confusion with the
previous document.

R=kinuko@chromium.org,engedy@chromium.org,csharrisson@chromium.org
BUG=998171,1061899,1068207

Change-Id: I69161c73224a5be4f57c6c3f5bca65042e341e50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2133976
Commit-Queue: Alexander Timin <altimin@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#767827}


CrOS-Libchrome-Original-Commit: 1cc31f46c2e6ae658a97b92fbc8c556eba382d3e