[2/4] Add ThreadControllerPowerMonitor to manage power states
This CL is introducing a helper class ThreadControllerPowerMonitor
that assist ThreadController for tasks scheduling.

The use of ThreadControllerPowerMonitor is optional and the
power state is assume to not be suspended when unused.
The end goal is to activate it by default after testing on stable
channel.

TBR=nasko@chromium.org

Bug: 1074332
Change-Id: I40c3580ccba0b4755efed19f264e4634e3f13770
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2200421
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771619}


CrOS-Libchrome-Original-Commit: 8e7026d226d87e606529e9c8f5962ab428167e66