ImportantFileWriter: Fix races part 5.
For base::ReplaceFile to succeed on Windows, both the source and target
files must have no open handles. Since ImportantFileWriter writes to a
temporary file, closes it, and then calls ReplaceFile with that
temporary file as the source, it will only succeed if no other processes
have an open handle on either file. On Windows, the very act of closing
a file that has been modified broadcasts a message to filesystem
observers that a change has happened in the directory holding the
file. Certain software may then try to open the file that was just
closed. This, of course, will block ReplaceFile. This CL boosts the
priority of the thread hosting an ImportantFileWriter for the brief
window containing the close and replace operations. With any luck, this
will help Chrome win the race against other software so that ReplaceFile
succeeds more often.

BUG=1075917
R=gab@chromium.org

Change-Id: I2dd286f63e8b269d209da7497167b53ea3a54ef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214624
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Auto-Submit: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771802}


CrOS-Libchrome-Original-Commit: 79a295e6d07b857e3257eab67d6a4b1c4c83085e