Fix multiple flakes in HangWatcher tests
Flakes were caused by a race between the first thread registering and
the first iteration of loop in HangWatcher::Run(). Problem is fixed by
always performing the Wait->Check->Possible Monitor cycle even if there
are no registered threads. The only caveat is that this means a
HangWatcher with zero registered threads will still wakeup on a
heartbeat to attempt monitoring. This is never the case post-startup in
practice.

HangWatcher::Start() now has to be invoked manually so it can be done
safely after invoking *ForTesting functions.

Bug: 1078828, 1081654
Change-Id: I52f2ea6bb66e6e05b36c52dbc3d8108e756055ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2196831
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Oliver Li <olivierli@google.com>
Cr-Commit-Position: refs/heads/master@{#771974}


CrOS-Libchrome-Original-Commit: 46db1739d516f5c8dbd2d4d36aff3a777d4352e9