Initial CheckedPtr2 implementation
This code is currently unused, but is needed to evaluate performance
impact of the CheckedPtr initiative.
https://docs.google.com/document/d/1pnnOAIz_DMWDI4oIOFoMAqLnf_MZ2GsrJNb_dbQ3ZBg

The current implementation is an expensive no-op. It offers no run-time
verification of pointer correctness, but it performs work as if it did.
It is expected to have similar performance characteristics to the
designed CheckedPtr2. CHECKED_PTR2_PROTECTION_ENABLED and "TEST"
comments show how the actual code would look like had all needed
support been already implemented.

Bug: 1073933
Change-Id: I4fc6a1a3b6701dcae9bb38ff3fa01fd6276812fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2162493
Commit-Queue: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772675}


CrOS-Libchrome-Original-Commit: 4ca50e35e8136b1166ec8a67304ad1ab17c357f1