Support launching the browser as a Mojo client
This adds support for launching the browser (for any Content
embedder) with a Mojo invitation on the command line, allowing
external clients to include the browser process in a pre-existing
Mojo process network and access a generic primordial Binder
interface akin to InterfaceProvider.

A simple browser test is added to verify that Content Shell can be
launched in this mode and the primordial Binder can be used by the
external launcher (in this case, content_browsertests) to bind a
new testing control interface (content.mojom.ShellController)
exposed by Content Shell.

See crbug.com/1082454 for motivating use cases.

Bug: 1082444, 1082454
Change-Id: I2cfc66d94aa9bdb5072e8ca63d95ae562b63b5b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2200004
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773320}


CrOS-Libchrome-Original-Commit: b86921c771bff056d2b018238b8749eb25535cb5