[4/4] Avoid running delayed tasks in power suspended state
This CL is suspending the execution of delayed task while the
system is suspended. The ThreadControllerPowerMonitor is
used to notify the ThreadController when the power is back to
normal.

The delayed tasks can no longer be executed between suspend and
resume notifications. They are postponed after the resume notification.

On Windows, while going to suspend mode or waking up, the computer has
bottleneck on clock API (system-wide lock) and has a lot of disk IO
contention. It is not the right moment to wake up all the pending
delayed tasks at once. It is better waiting for the resume notification
while the computer is in a running state.

Bug: 1074332
Change-Id: Icda16b76fc87a20fee22997baf7eeb5e5d02eb09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2186200
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773354}


CrOS-Libchrome-Original-Commit: 6250a0340f5b8665b37c9bd71ee3f083c53fa8c8