[Sampling profiler] Avoid metadata read lock assertion
The assertion is implemented with DCHECK so is non-reentrant. It cannot
be safely invoked from the GetItems() functions since they are invoked
while the target thread is suspended. Independent of that the assertion
does not work in general on Android because the checking done by
base::Lock::AssertAcquired() is too restrictive.

Bug: 1004855
Change-Id: Ia2b390283844adf640077e7ee34650bf610cca09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225784
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#774841}


CrOS-Libchrome-Original-Commit: cdb207e6c45aeeb331d3d50957edad33de888f22