android: Report child service exception during init
Exceptions before crash reporting is enabled in child process is not
reported, and are treated in metrics as android kill.

Catch these exceptions in the child process and send the exception
string to the browser process. Report these as java exceptions for crash
reporting. And plumb the information all the way to
ChildExitObserver::TerminationInfo to make sure metrics record this as a
crash.

Bug: 1087966
Change-Id: I9599b31cb0338201e4c87c955a2dbc4ae55c4466
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2229758
Reviewed-by: Peter Wen <wnwen@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#775346}


CrOS-Libchrome-Original-Commit: dd286b35953c5dde66d20b0ca57b80d1a343b084