Support policy driven search provider context menu search access
Added new policy called DefaultSearchProviderAllowContextMenuAccess.
Provides a way to enable/disable the "Search <Search Provider> For"
context menu item present when a text selection is highlighted in
a page. The policy value is only appled when the
"DefaultSearchProviderEnabled" policy is enabled, and is not
applicable otherwise.

Bug: 985951
Change-Id: Iecf4da8793a1655bff6b9401b85a880a018c253f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225350
Commit-Queue: Roman Arora <romanarora@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777142}


CrOS-Libchrome-Original-Commit: ebf0ed7f2c7260bcb4eecd8b790104d58b139846