Clean up non-inclusive language in //mojo and //ipc
Replaces use of the term "master" with "primary" in the context of
associated interfaces, referring to the primary interface which owns and
operates on a pipe.

Also replaces the only other use (a reference to relative importance of
a process within a hierarchy) with "root", which is both more inclusive
and more appropriate in context.

Bug: 1097211
Change-Id: I0bd43e093b20e61bc4f15c34d2bcb82867705975
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255206
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780538}


CrOS-Libchrome-Original-Commit: cd23f75e19dcf4a9ae2e3e01f83cbcb3044d007e