Remove/replace unnecessary includes of logging.h (ui/)
If the file just needs the CHECK/CHECK_OP/NOTREACHED
macros, use the appropriate header for that instead.
Or if logging.h is not needed at all, remove it.

This is both a nice cleanup (logging.h is a big header,
and including it unnecessarily has compile-time costs),
and part of the final step towards making logging.h no
longer include check.h and the others.

TBR=kinuko

Bug: 1031540
Change-Id: I970f0010731c6190f3fd73e3f4643e1369442ef4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254326
Commit-Queue: Hans Wennborg <hans@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#781228}


CrOS-Libchrome-Original-Commit: 4b46b3b9ccf838e90dc05a81584ae1118af97ce8