base/synchronization: Always Try() in LockImpl::Lock().
The current implementation of base::Lock on Windows and POSIX assumes
that base::LockImpl::Try() performs a kernel call, which is not true on
Linux, macOS and Windows at least, since they all use a variant of "fast
user-space mutexes" (e.g. futex() on Linux).

Since these calls are cheap (On Android for instance, an external call
to libc and a compare-and-exchange with acquire ordering), do them all
the time. An alternative is to remove the branch altogether, as
base::debug::ScopedLockAcquireActivity checks whether it is enabled as
well. However it still does a non-zero amount of work, so we may as well
avoid it, assuming that locks are mostly uncontended.

All in all, assuming that the lock is not held and that ActivityTracker
is not enabled, this replaces:
- 2 acquire loads (for the if, and then in ScopedLockAcquireActivity)
- A predictable branch
- A call to to Try()

with a single call to Try() (as the native lock() call will effectively
try() first).

Change-Id: Ie77392db4008966b6379eaadb5d20d1ebfcbc276
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270203
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784736}


CrOS-Libchrome-Original-Commit: 6cabb581eb80a101f9365c12beaa38c1df7b46c9