Avoid JSONPerfTest OOM failures on x86 builds
JSONPerfTest.StressTest was crashing on x86 builds. Debugging showed
that it was hitting an OOM failure and analysis showed that this was
because some iterations of the test required more than 4 GB of memory.
No memory leaks were seen.

This lowers the maximum depth from 12 to 10 which drops the memory
required from ~6 GB to ~250 MB, and the time from 58 seconds to 4.
This should be enough to avoid Android timeouts so this change also
re-enables the test for Android.

Bug: 906686
Change-Id: I1114159da9ffb1a0694810d5686bba42ca671327
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283966
Commit-Queue: Albert J. Wong <ajwong@chromium.org>
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Albert J. Wong <ajwong@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786176}


CrOS-Libchrome-Original-Commit: bf3bd23a4bae026cfc0b49fe2e4d9743e4a16131