Implement basic screen-sharing for lacros.
This CL creates a new mojo interface: ScreenManager. This interface has
a single method TakeScreenSnapshot which returns a bitmap of the
screen's contents. Lacros-chrome uses this interface to implement a
rudimentary version of screen-sharing. This in turn allows
meet.google.com screen sharing to function properly.

This interface will eventually be supplemented with methods to allow
window-sharing as well. The whole interface will likely be overhauled in
the future to create a more performant and fully-featured version of
screen/window sharing.

Change-Id: I821c59464ea4d2ff57665381038837f29645750d
Bug: 1094460
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2261430
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Auto-Submit: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786507}


CrOS-Libchrome-Original-Commit: 438f05415e9888c0de66f7f05a00724241127514