chrome/android: Record battery capacity drain into UMA
Adds AndroidBatteryMetrics to chrome/android, which track changes in
remaining battery capacity levels while Chrome is in foreground and
the device is on battery power. The deltas are reported into a
cumulative UMA histogram. so that we can reconstruct per-user daily
battery drain in the dashboard. We also report per-30-second drain
values to facilitate using battery drain as a trigger for field tracing.

Test: Run Chrome locally (Pixel 4), check chrome://histograms.
Bug: 1102048
Change-Id: I77aa38d9926b155633a8e6d81f46062ff26dccbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2281181
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787200}


CrOS-Libchrome-Original-Commit: 3c1f99e36b53c9a51c0af4f7b45bd524a1f36e79