Migrate Omnibox native java unittests to batched Instrumentation tests
Native java unittests are going away in favor of batched instrumentation
tests. This CL migrates the Omnibox tests.

Note that I had to make changes to the tests that used the clipboard
service, as these tests run without a focused Activity, and so calls to
the real clipboard service are rejected.

Bug: 1103344, 1102026
Change-Id: Iec82e26ce28e503944dff82027f3a3e083ab7742
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2288379
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Sam Maier <smaier@chromium.org>
Reviewed-by: Ender <ender@google.com>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790162}


CrOS-Libchrome-Original-Commit: b15152b2b49be75fc127ff2b2b64d858b3102181