MojoLPM: Simplify the protobuf format for Remote/Receiver types.
This change simplifies the protobuf representation of Remote/Receiver
types, and adjusts the generated code accordingly. It also removes some
duplication of code in the Remote/AssociatedRemote cases so that the
handling of calls to remotes need only be implemented once for each
interface.

Bug: 1076336
Change-Id: Ib4cdb0cb6cf700ebb74feb28cd66acb3b49f2d30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302609
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Oliver Chang <ochang@chromium.org>
Commit-Queue: Mark Brand <markbrand@google.com>
Cr-Commit-Position: refs/heads/master@{#792256}


CrOS-Libchrome-Original-Commit: ab8018a7d99037cd93a6de1dba8c9d4210dcaff6