Only HangWatchScopes hung at time of hang detection should block.
If a hang is detected in a thread critical to user responsiveness like
browser UI or IO thread then blocking execution in ~HangWatchScope() on
all threads makes sense because the program was already unresponsive.

It's desirable to keep the blocking to a minimum on the browser UI/IO
threads if they are not the hung thread. This CL achieves that
by only enforcing blocking in HangWatchScopes that were hung at the
moment of hang detection.

Bug: 1034046
Change-Id: I05aace41e272f1daef490be55902225363536396
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2209330
Commit-Queue: Oliver Li <olivierli@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792452}


CrOS-Libchrome-Original-Commit: 32b5eff94581a1d0d76023ffe9c08e9aadfa9700