Rename TimeDelta::operator/(TimeDelta) to IntDiv().
No functional change.

This is temporary; the ultimate goal is to reintroduce operator/() as
doing floating-point division, and adjust consumers accordingly.  For
now, this (a) makes uses obvious/greppable/cross-referenceable, and
(b) provides a name that will be distinct and clear against a
floating-point "div".

Bug: 1104532
Change-Id: I2e55f717648d3cdcbe4ad6342bab92f16e926b97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316600
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793926}


CrOS-Libchrome-Original-Commit: d0bcff98cd11c999c42aae0be9e2ab122b791015