[base] Make PartitionAlloc everywhere work on Android.
On Android, PartitionAlloc everywhere doesn't work, because not all
callers are properly wrapped. This is causing issues in at least one
case, when a caller uses UncheckedMalloc() to allocate, and free() to
release the memory. The allocation is intercepted, but not the
deallocation, making Chrome crash.

The proper solution is to make sure that all callers are properly
intercepted, but even with that, PartitionAlloc's behavior with
UncheckedAlloc() is not the intended one. Calling malloc() at least
documents this behavior, as it is configured in Chrome to crash in case
of allocation failure.

This is tracked in the linked bugs, but at least this commits addresses
the correctness issue.

Bug: 1111331, 1111332, 998048
Change-Id: Ie9131db513a502bec3b3ceb6ac52fa8b8dca9b95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2329850
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794041}


CrOS-Libchrome-Original-Commit: 2b7a74647f903c10dcd4189be5ad80a716d3e568