Remove deprecated RangeF rounding APIs
The rounding APIs on a Range are not used anymore.

We fixed the use cases where the call to Round was incorrect.
Typically, the code was looking for an "enclosing-range" or
and "enclosed-range".

It is preferable to remove the API since it's misleading the devs.

R=rsesek@chromium.org

Bug: 1111044
Change-Id: I7f11c9c28c2a5c93d3fe8cfcbd38a6b077e90105
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339713
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795570}


CrOS-Libchrome-Original-Commit: 68239f6a9c4df8fb1265f8311a40a0fdc1b99355