chrome/android: Add experiment to run all execution on little cores
Adds a feature-list-controlled experiment to restrict execution of all
Chrome processes/threads to LITTLE-cores only (on big.LITTLE-like
architectures).

We intend to run a limited experiment on Canary/Dev to estimate the
potential power savings of running on the more efficient little cores,
as well as the impact of this on system health metrics.

We also plan to add a similar experiment for WebView. However, since
WebView shares a process with the hosting app, we may need to be more
careful about which threads we opt into little-core-only execution.

TBR: dcheng@chromium.org
Bug: 1111789
Change-Id: If2bc8c811692f7e716428398db3e5d0649da42d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332715
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796342}


CrOS-Libchrome-Original-Commit: a0467842112bb57115e1432c64041ac706db48f7