[WebLayer] Support building WebLayer as an isolated split in WebView
This adds basic support for building WebLayer as an isolated split in
WebView. A new gn arg "weblayer_in_split" is added which will move
WebLayer Java code to it's own DFM if true.

Supporting isolated splits required some client changes in WebLayer, so
I added a check in WebLayerFactoryImpl.isClientSupported() to make sure
isolated splits aren't used with an unsupported client.

This also adds the ability to write the version_name and version_code of
a bundle module to the build_config, which allows DFMs to match the base
module version info (an error is thrown if these don't match). This
makes it easy to share version info, since computing WebView
version_code is pretty complicated.

The bundle with isolated splits won't install correctly with the current
version of bundletool we have in Chromium, but I'm working in parallel
on fixing bundletool.

Bug: 1105096
Change-Id: Id5522b42e283dc813fa84aaedc806f7519b586ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2353378
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798815}


CrOS-Libchrome-Original-Commit: c5f986b96cff358525864463b5fce0c195f6b9a2