Delete disable-component-cloud-policy switch
This flag is introduced with https://codereview.chromium.org/250383004
as a way to disable the feature. Give that the feature has been launched
for many years, we probably don't need this anymore.

And it gives user a way to escape component cloud policy on desktop.

Bug: 1119515
Change-Id: Icefcdc2ea49d9b25dbb535616befef427a53ee90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2364960
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801844}


CrOS-Libchrome-Original-Commit: fa7dd9bbccd181d02a494bd26bf18dc34cf7f229