[base] Make RunOnceCallback support move-only types
This change adds support for move-only types to
base::test::RunOnceCallback. If a move-only type is passed to the
constructor of RunOnceCallback, it will later be passed by move to
the OnceCallback that should be run. This invalidates the action,
and running it twice will result in a run-time crash.

Bug: 752720
Change-Id: I43ca5e81e9323c7c374e67de14be2637fbfea589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2379840
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#803630}


CrOS-Libchrome-Original-Commit: b2df556a29a5288545c6cc8908a9fe558bc817c6