Let blink runtime features override the default status of chromium features
This honors blink feature status ("test", "experimental", "stable")
in third_party/blink/renderer/platform/runtime_enabled_features.json5,
overriding of the default status of the corresponding chromium features,
unless the status of the chromium feature is overridden by finch trial
or command line switches.

This resolves the following problems of blink features with
status:"test", status"experimental", status:"stable" when the
corresponding chromium feature is disabled:
1. unexpected disabling of the blink feature for web tests
2. unexpected enabling of a blink feature for unit tests (in which case
   the blink feature should not have a status).

This CL only affects how blink features with status are enabled with
--enable-experimental-web-platform-features
(or chrome://flags/#enable-experimental-web-platform-features)
--enable-blink-test-features when the corresponding chromium feature is
disabled by default.

This doesn't affect
1. finch experiments (which still controls the blink features through
   the chromium features);
2. other command line flags that affect blink features (either directly
   or indirectly through chromium features);
3. the priority relationship among 1 and 2 and the changed factors
   mentioned above.

The features that will be affected by this CL have been prepared in
either of the following methods (tracked by crbug.com/1121839, based
on feedbacks in https://docs.google.com/spreadsheets/d/1DAtdbGy7XZCC_1_IXHXZKx568pw0eDlVETCtJBexqZc/edit#gid=0):
a) For features that should be enabled for web tests but were not
   because of the chromium feature disabled by default, temporarily
   explicitly enable it in content/child/runtime_features.cc.
   An example is crrev.com/c/2380213.
   These changes are removed in this CL.
b) For features that are not ready to be enabled for web tests, or are
   not meaningful to be enabled for web tests (e.g. the feature is
   implemented at both chromium side and blink side), the blink status
   is removed to let the chromium feature fully control the blink
   feature. An example is crrev.com/c/2376067.

Bug: 1048656, 1121839
Change-Id: I8b2f15e30767dce9212097f79c8c07a373af892d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2375886
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Jason Chase <chasej@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#803712}


CrOS-Libchrome-Original-Commit: 05355f4ad601c0daa6d51c5ac0d035c06e148e44