[base] Remove markers around on-stack copy of fatal log messages.
LOG(FATAL) message content is now provided in Crashpad dumps via an
annotation, so there is no longer value in providing markers to allow
the stack to be scanned for the message.

Removing this also avoids the need to update the test not to cause
failures in ASAN builds.

Bug: 802393, 1119821
Change-Id: I6a64778f4da527597e9bb0874fe5764fbf87716a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2404926
Auto-Submit: Wez <wez@chromium.org>
Reviewed-by: Joshua Peraza <jperaza@chromium.org>
Commit-Queue: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806252}


CrOS-Libchrome-Original-Commit: d78fee6dc9bd50c75f3b250ca4ac482053dec673