Cancel augmented Autofill session when URL bar loses focus
Since Android Q, views that are labeled with importantForAutofill=no
can still trigger the augmented Autofill service. If this happens, no
ordinary Autofill requests will be triggered anymore for as long as the
Autofill session remains active. This prevents the user-configured
Autofill service from running on web contents if the user's first
interaction is with the URL bar.

This issue is resolved by canceling the current Autofill session when
the URL bar loses focus.

R=dmazzoni@chromium.org, michaelbai@chromium.org

Bug: 1103555
Change-Id: If06c15f4c397107c284c4de757f58c3aeb939cf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302636
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808524}


CrOS-Libchrome-Original-Commit: 444aad9599a465e3db4c52b3331c5c286818e8bc