Remove ConvertPointToPixel() methods.
The Convert*ToDIP() and Convert*ToPixel() functions are problematic in
that they hide what decision is being taken in the underlying code
when moving from floating point to integer, and consumers should be
thinking about what they want for correctness. They also act in
inconsistent ways when converting a Rect vs its components. This
removes the functions that perform float->int conversions in order to
have callers show what they intend to happen and make it clear that
data is being lost in some fashion.

R=sky@chromium.org

Bug: 1130050
Change-Id: I2bbe5ff2b85cc4c643ec66fe09a8208e643c3329
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2418568
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#809815}


CrOS-Libchrome-Original-Commit: 8c8c4211424894e79c1441f9fcd90dc8effd02e2