content: Enforce CPU affinity on Android
As a part of the experiment to run all execution on little cores
we want to make sure that the cpu affinity doesn't change at runtime,
e.g. when Chrome goes back from background. To do this, we introduce
an interface in content that allows to enforce affinity by checking
it on a regular basis (every 100 tasks) and setting it back when it
changes.

Change-Id: I766e94af42861870abe3a0079dd335a23462d481
Bug: 1111789
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2418673
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812027}


CrOS-Libchrome-Original-Commit: a2483f8c25a8d70d4dac61a051ec211ce10e887b