Remove implicit-conversion-to-integer ConvertRectToPixel() methods.
The Convert*ToDIP() and Convert*ToPixel() functions are problematic in
that they hide what decision is being taken in the underlying code
when moving from floating point to integer, and consumers should be
thinking about what they want for correctness. They also act in
inconsistent ways when converting a Rect vs its components. This
removes the functions that perform float->int conversions in order to
have callers show what they intend to happen and make it clear that
data is being lost in some fashion.

R=sky@chromium.org

Bug: 1130050
Change-Id: I042735e7cc761da1a581ddb7b909bbedf4bd86be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2429143
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812168}


CrOS-Libchrome-Original-Commit: 340909edba6daccd34d5875de93599551b218902