Modernize scaling of gfx::Insets
The current Insets::Scale() methods scale and then do a static_cast to
integers. This does not provide the same functionality that now exists
for other geometry types.

We add ScaleToFlooredInsets(), ScaleToCeiledInsets() and
ScaleToRoundedInsets() to replace the Insets::Scale() method when
scaling from integers to integers, to make the operation being done
explicit.

This also means we provide explicit conversions with ToFlooredInsets(),
ToCeiledInsets(), and ToRoundedInsets().

We also add ScaleInsets() for InsetsF to allow for floating point
scaling.

We update the ConvertInsetsToDips/Pixels helper methods to use the
new ScaleInsets() method, and return an InsetsF, similar to other
conversion helpers.

And we update callers to Insets::Scale() to either use the device
scale helpers and ToFlooredInsets(), etc, or to use the
ScaleToFlooredInsets() and friends.

Unit tests are updated and added for the new scaling methods.

R=sky@chromium.org

Bug: 1130050
Change-Id: I86dbe8b4dd43b2fd18521c07f251178ded9c5dec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2429383
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812385}


CrOS-Libchrome-Original-Commit: 0eff6293450117295167e6aafbbfb3ee9a80ad30