Adds mojo_base.mojom.ReadOnlyFile
Adds a mojo ReadOnlyFile that won't serialize if it can be written.

The expectation is that this handle will then be sent to another process
which lacks the rights to add any write permissions to the handle. This
may not be true on all platforms but is useful where this holds.

Code to check for readonly-ness is implemented within mojo rather than
base::File as it is (a) difficult to say in general what it means to be
readonly, and (b) not possible on every platform.

Bug: 1130762
Change-Id: I6f00934005da7e433af4fee94454dfcd28dfd3c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2417549
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Cr-Commit-Position: refs/heads/master@{#813987}


CrOS-Libchrome-Original-Commit: a358e673f879c77dbb23fc4d1017485b53e58705