[Bluetooth] Add ScanResponseData to BLE Adv D-Bus API
This is a temporary change to support scan response data in BLE
advertisements for Nearby Share. Once the final API lands in bluez
upstream we will need to revert this change (tracked here:
https://crbug.com/1136920).

This allows a map of byte (BLE type) to byte array (payload) to be
supplied as the scan response for a BLE advertisement. Currently only a
single type 0x16 is allowed in the map by bluez.

The corresponding bluez change is here: https://crrev.com/c/2458306.

A follow up CL will supply the scan response for Nearby Share.

Bug: 1135699
Change-Id: I3e2c8098a7cfca2025f0c7cd5c6b631156e77de0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2459250
Commit-Queue: James Vecore <vecore@google.com>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#816037}


CrOS-Libchrome-Original-Commit: 85b2daf9e552c32e10f0a9137bec92856fa02d60