Mojo: Fix deadlock in quota control messaging
If an intraprocess pipe calls SetQuota on one endpoint, it's possible
for this to result in a deadlock due to Mojo internals re-entering the
endpoint's MessagePipeDispatcher while already holding its internal
lock.

This fixes the issue about deferring the potential re-entry until the
lock is released.

Fixed: 1137988
Change-Id: I814415a3ce155a991a2e6d2cfd07fa309004eeda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2485147
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#819221}


CrOS-Libchrome-Original-Commit: 36cbabe8f5addb05533065957a560321e43cde8c