[scheduling] MainThreadTaskQueue has a TaskQueue instead of extending TaskQueue.
This cl modifies MainThreadTaskQueue so that it owns a TaskQueue instead
of extending the base/sequence_manager/TaskQueue. This is the first
step in a TaskQueue ownership refactor described in this doc[1].

For now, MainThreadTaskQueue remains reference counted but this will
change in follow-up CLs and these queues will be uniquely owned instead.

[1] https://docs.google.com/document/d/1E--SK7wIuxwd6aGNiPc0uEU8jm1tvAvD1UrHyipn9sQ/edit?usp=sharing

Change-Id: I124103bde681fe1f55c87b701691b53116f626a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2453591
Commit-Queue: Katie Dillon <kdillon@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#821307}


CrOS-Libchrome-Original-Commit: 993f9b50f3733f094f455a6c4879fb16798f746f