Get rid of ViewMsg_MoveOrResizeStarted message.
This message which is only sent on windows and ash. Looking back on the
history (http://crrev.com/30f75e61) when this message was added it was
intended to hide auto complete popups. Being that we dismiss select
popups on focus changes and size changes I believe this method is not
needed anymore.

BUG=1051648

Change-Id: I463da85c9916c2140e4093447df1aacefc75b9c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2493282
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#821841}


CrOS-Libchrome-Original-Commit: b8a4a089b2dd3a4190a8d33bbec74257cb35d4c9