[base] Make sure 'ThreadController active' trace event outscopes all other RunTask events
This was missed in the original implementation because on Windows
the pump always processes a native event before DoWork() and
hence always kicks off the "ThreadController active" trace event
prior to DoWork(). On platforms that don't do this, the tasks nested
backwards (and the ends were out of order which messed up the trace
layout).

Also cover DidRunTask() now to cover microtasks and added extra brackets
to make it extra clear what the scope of "things we do to run the task"
are.

R=altimin@chromium.org

Bug: 899897, 1074019
Change-Id: I67a01853365e173ed924513592191d6098a4f621
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2536998
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#827482}


CrOS-Libchrome-Original-Commit: be1a4e4c5bfe54e17f052d740841f6324d51f85f