Require SSE3 on x86 platforms.
Before this CL, Chrome was built assuming SSE2 support on Windows,
Linux, Chrome OS, and Android. This CL raises the assumption to SSE3 on
all these platforms. This CL is a no-op for MSVC builds, which don't
have a way to explicitly target SSE3.

This CL does not impact Mac, where we only support 64-bit builds, which
allows us to assume SSSE3 support. (The first 64-bit Intel Mac used the
Core 2 microarchitecture.)

This CL will not have a usage impact on Chrome OS, which already assumes
SSE3 support, and on Android, where we require Lollipop+, which is only
used in phones with SSSE3+ support.

This CL will have a tiny impact on Windows and Linux usage. The impact
has been quantified and deemed acceptable. Impacted users are receiving
a notification starting in M87.

Confusingly, "SSSE3" is not a typo of "SSE3". It is the next step up
after SSE3, and before SSE4. This CL will cause Chrome to require SSE3.
Targeting SSSE3 is future work.

More details can be found in the design doc linked in the attached bug.

Bug: 1123353
Change-Id: If93aa9d37e75f3068b06c42f7412c7c4578f3b9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2311044
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#827575}


CrOS-Libchrome-Original-Commit: 5bb2864fdd57e45c84459520234b37a01e7a015a