Mojo: NodeChannel should require a minimum size when using multiple versions
When we added support for versioned message we accidentally introduced a
behavior where messages can be shorter than expected. This is what the
behavior should be to allow older messages to be received. However,
we know the minimum size is the oldest message version. We should
never be allowed to go smaller than that.

BUG=chromium:1149805

Change-Id: I3a24dd62b6fafd742cd7ac155f863d33bb5c6de8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2545207
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Brian Geffon <bgeffon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#828478}


CrOS-Libchrome-Original-Commit: 8ad25003de8082128c3b34565ea288dc78fa88db