Fix OverlayWindowViews sizing for large and small aspect ratios
This fixes the following issues with OverlayWindowViews sizing:

 - Windows: For large and small aspect ratios (e.g., 10:1 on my display)
   resizing the window was badly broken, with the window disappearing
   from the work area.

 - macOS: The minimum size was not enforced properly. For example, it
   was possible to resize a square window down to 146x146 pixels, which
   ignores the 260 lower bound on width.

To fix the Windows issue, SizeMinMaxToAspectRatio() is merged into
SizeRectToAspectRatio() and protection is introduced against exceeding
the size limits. So for example, even for a large aspect ratio and a
certain maximum width, SizeRectToAspectRatio() will never produce a
height that is below the minimum height specified.

SizeRectToAspectRatio() is now also re-used on macOS as a replacement
for the Cocoa implementation of aspect ratio enforcement, which suffers
from the issue described above.

Bug: 1123249
Change-Id: Icdd5bcd920d940d7d46926a199300ce305c7f430
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2552394
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Wojciech Dzierżanowski <wdzierzanowski@opera.com>
Cr-Commit-Position: refs/heads/master@{#834479}


CrOS-Libchrome-Original-Commit: 8a3927cdbddfd2a03a94990138235f0acfed4d91