Limit AImageReader max size to 1.
1. Limit AImageReader max size to 1 based on the model name of the device.
Model name of the device is sent from server side using the finch
experiment.

2. There will be an equivalent critique CL to add a list of known
devices which can only support max size of 1. This CL will be checked
in first.

3. Move the finch experiment from media/ to gpu/ to avoid dependency
issues.

4. Add error logs to print device model name for the devices which fails
creating image reader with max size > 1. This is required so that the
model name can then be added to the finch experiment.

Bug: 1051705
Change-Id: Ie4809da3591b0535ca129642f248bf9534e37874
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2578385
Commit-Queue: vikas soni <vikassoni@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836751}


CrOS-Libchrome-Original-Commit: 21c0ee07d7d7667f079f46b79a75889afb8661ee