Lacros: Make condition of is_chromeos_lacros tighter.
This change makes is_chromeos_lacros to be true only when current_os
is either linux or chromeos. This ensures that is_chromeos_lacros will
evaluate to false when generating ninja files for toolchains such as the
toolchain for pnacl.

It also ensure that enable_scrambled_message_ids is false for
lacros-chrome regardless of toolchain that is in use.

BUG: 1052397
Change-Id: Ibbe8d2149937b506c6c86f2422a356c5d2d4d869
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2581632
Commit-Queue: Yuta Hijikata <ythjkt@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#837409}


CrOS-Libchrome-Original-Commit: a6136cb16652b7b3eaa5d6d5d80296fe4b1b9509