Add 8B to all allocations when Win x86 binary is run on ARM64
This change (based on crrev.com/c/2576881) is a temporary workaround for
a bug in the x86 emulator on ARM64 Windows. The bug is that movvdup
reads sixteen bytes instead of eight and this makes Chrome unusably
crashy. The fix is to add eight bytes to allocations so that the
over-reads aren't fatal.

This change will make x86 Chrome on ARM64 Windows use slightly more
memory. On normal x86 builds the only difference should be an extra add
in the allocation path.

The emulator bug has been reported and fixed. This change will be
reverted some time after the fix has shipped. See the bug for details.

This has been manually tested on ARM64 hardware to confirm that the
crash is avoided with this change. This has been more extensively tested
with a modified version of this change (crrev.com/c/2590684) which
always adds the eight bytes.

This change will be reverted when the emulator fix has shipped.

Bug: 1151455
Change-Id: I2fcd9fb5169239b9957e9e8838f441f5780e2610
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2577893
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#837901}


CrOS-Libchrome-Original-Commit: c54407621ee9023f14bdc449fc2b6b6d03f8907b