Record histogram for app zygote start time
This adds a histogram for the time it takes to start up the app zygote.
This is similar to the other Android.ChildProcessStartTimeV2 histograms.
To do this, the zygote saves its startup time in doPreload, and we then
send this time back from a renderer if it was started from the zygote.
The browser records this time once per startup.

Bug: 1150600
Change-Id: I2533e68ea40f417d337bc76c76d90bdd976d2e15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2597765
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#838339}


CrOS-Libchrome-Original-Commit: d023f37a21f2405d15ddce8d3a9dfc1264af5a26