Add FetchCapabilities to print backend service
Incorporate support for retrieving a printer's basic info, user defined
papers, and semantic capabilities and defaults from the utility service.
This combined collection of information is needed when preparing for
print preview, so it reduces overhead and latency to enable these to all
be collected with one mojom call instead of needing a chain of calls to
assemble the data.

Bug: 809738
Change-Id: I75518dca30f0113cadc7971ba2a87658aca7176b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2578698
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#841347}


CrOS-Libchrome-Original-Commit: c15aea6791f4a6c4e79903cf60819138d4a6f5b2