Fix ClassLoader mismatch for modules which depend on the chrome module
I noticed most of the exceptions when loading the vr module have logs
from SplitCompatAppComponentFactory, which means the chrome ClassLoader
is messed up in the ClassLoader cache. When the vr module is loaded, it
uses the wrong chrome ClassLoader as the vr ClassLoader parent, which
ends up causing a ClassCastException. This CL detects that case and
creates a fresh ClassLoader with the correct parent.

I tested this locally by faking the situation where the chrome
ClassLoader is incorrect, and this loads the vr module and other DFMs
which depend on chrome (e.g. image_editor) without crashing.

Bug: 1146745
Change-Id: I5960fa2817604a13b90d80b10517b3b416c46013
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2620939
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#842167}


CrOS-Libchrome-Original-Commit: 77c454fbe247891187ec96b422f23be8cb2d742d