base/android: ModernLinker: do not wait for RELRO to arrive
The LegacyLinker still wants to wait until RELRO arrives (because there
is no atomic RELRO replacement support yet - but it is not hard).

In order to accommodate both waiting and delayed use, the Linker.java
looks whether the subclass waited for RELRO, and if there was no
explicit waiting, assumes the atomic replacement is implemented. This
parent/subclass coordination is a bit unfortunate, as it makes the state
transitions hard to understand, but the implementation of it is small.

I do not expect this to have a big impact on startup. My local testing
did not reveal any evidence of faster renderer load time.

Potentially in the future mLibInfo could be guarded by another lock, so
that it could arrive *while* the native library is loading. Currently it
is not done, as it would require a larger change. I think it would have
a chance to help only if this change as-is provides some nontrivial
help.

New metrics are introduced to try to avoid the overhead of a fair A/B
test. They would approximate how many times a cold browser process makes
a shared RELRO region such that it arrives *after* the child has loaded
the native library (in other words, the child is not blocked). These
metrics do not account for being in background and such, so it is a very
rough approximation.

Lab performance testing summary:

android-go-perf,android-pixel2-perf,android-nexus5x-perf: no change

Bug: 1154224
Change-Id: I6d385280910424e1d6b09784650d72b598a1ba23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2610030
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#843584}


CrOS-Libchrome-Original-Commit: abcbc6710b5c634917a7c9f724ecf4282e303329