[CrOS Settings] Add support for showing Month year in UTC.
The EOL dates listed at
https://support.google.com/chrome/a/answer/6220366?hl=en
do not provide a timezone.

However, they refer to UTC instead of a local time zone. (End of life
occurs at a single point in time).

Since users are NOT told that the time zone is in UTC, they will likely
confuse the end of life date to be in their local timezone.

If the end of life is on the first day of the month, the end of life
will show as one month sooner than the actual end of life.

This CL adds an extra parameter to base::TimeFormatMonthAndYear()
to allow for time formatting in UTC instead of the local time.

Fixed: 1126296
Change-Id: Iede26508a9a30e776225095bd6267bd2c2fe64eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2399826
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844088}


CrOS-Libchrome-Original-Commit: d502ad61fb45711939d7213a458ece813eb0bd86