[scheduler] Add throttler methods to MainThreadTaskQueue.
This CL is a follow-up to this CL[1]. It adds some methods to
MainThreadTaskQueue such that anywhere where we need to pass a
pointer to the task queue when doing thorottling logic happens on
the MainThreadTaskQueue. This change moves us towards removing access to
the raw pointer to the task queue outside of MainThreadTaskQueue.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2453591

Bug: 1143007
Change-Id: Ie5f1ace08cc3329a89e5784010cc936d00832499
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2521443
Commit-Queue: Katie Dillon <kdillon@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845922}


CrOS-Libchrome-Original-Commit: 465f56090c505057625862d06fdcf4a5822a1f2a