[CodeHealth] Ban new usages of base::CallbackList
base::Bind and base::Callback are being deprecated in favor of the
{Once,Repeating} versions, and the same goes for base::CallbackList;
this adds a new PRESUBMIT check to ban new usages of base::CallbackList,
while still allowing existing usages.

base::ClosureList is similar in that it should be disallowed, but it was
not used anywhere. As such, the alias has been removed and replaced with
a comment that it should not be re-added.

Bug: 1113007
Change-Id: I1cc60d9b60321c232352d2eb1c1f9df7a48c8c08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2657022
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#848795}


CrOS-Libchrome-Original-Commit: 46a9271976d946b24c8e68724312e316de5105d5