[BackupRefPtr] Add reference count slot cookie
The majority of BRP crashes we've seen so far occur when a CheckedPtr
is assigned:
 - a random value that comes from an uninitialized memory
 - an already freed allocation
 - an allocation that has a reference count mismatch due to, e.g.,
   incorrect reinterprect_cast<> usage
 - an address outside the initial allocation slot via |operator+=|

Add a cookie to the reference count slot that helps us ensure that
slot pointer calculation is correct and that slot itself is not in
the freed state. The change should make diagnosing all three above
cases much easier.

Bug: 1073933,1165613
Change-Id: I32892f55148908a9ad14db4ecfe29300f53cb766
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2672579
Commit-Queue: Sergei Glazunov <glazunov@google.com>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850681}


CrOS-Libchrome-Original-Commit: 1e3e59a1d61b31ee0132383ac6194b0473a6f62e