[PA][BRP] Adjust pointer before checking belonging to GigaCage
CheckedPtr may legally contain pointers right past the end of an
allocation and we must be careful to not erroneously attribute them to
another slot.

PartitionAllocGetSlotStart() had a logic to subtract ref count and
cookie size from the pointer. That wasn't enough, because we shouldn't
get on this path to begin with if the pointer is outside of GigaCage
(see linked bug), so the adjustment has to happen when checking for
GigaCage belonging too.

Side note, there is no need to subtract that much, subtracting 1 will
suffice and result in a smaller instruction. We care only about pointers
literally at the end of an allocation, which is a common programming
pattern, e.g.
  T a[SIZE];
  CheckedPtr<T> p = a;
  p += SIZE;

Bug: 1169579
Change-Id: Ic0a6facd78f6e58d182fa2cb80f30226b4470f14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2651774
Commit-Queue: Bartek Nowierski <bartekn@chromium.org>
Auto-Submit: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850705}


CrOS-Libchrome-Original-Commit: 4d35d0f6f46e9c9b86b4a0a43334651d33b81bb2