atomicops: Retire Release_Load
Release_Load is a legacy atomic load with a Release constraint;
it implied a constraint that is not expressible in the standard
C++11 memory model. As of crrev 2680949, it will have no more
users; its partner Acquire_Store was retired in commit 77ab8e5f.

Bug: 420970
Change-Id: I084270df050b68273748589ca91201bf204080b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2670192
Commit-Queue: Venkatesh Srinivas <venkateshs@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Alexander Potapenko <glider@chromium.org>
Cr-Commit-Position: refs/heads/master@{#852892}


CrOS-Libchrome-Original-Commit: 879ddd5db6ec7a212d352bda2ef7a5b79c161525