Ensure we don't revert a preallocation
The ImportantFileWriter, which uses this class, preallocates memory for
the serialized JSON string. This CL ensures that we don't accidentally
reset this to a smaller value again. Implementers of the standard
library are free in their design to use the reserve() hint to allocate
less memory.

Bug: 1132866, 1173957
Change-Id: I61162eb80a40061c342b8811cf08617ae1986d8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2684279
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Auto-Submit: Dominic Battré <battre@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#853709}


CrOS-Libchrome-Original-Commit: c75c5852303a9fedaf654bad1d90cb68d5facda2