Add SCHEMA_ALLOW_INVALID_LIST_ENTRY and use it for DLP policy.
The new schema validation strategy allows unknown elements of lists,
just dropping them if present.
This allows list policies to be enhanced with more possible value,
but do not break the policy on the old clients.

DataLeakPrevention policy is the first to use to allow adding more
restrictions and levels in the future.
The unsupported ones will be just dropped on the older clients,
while the valid still be applied.

Bug: 1182281
Test: Updated unit tests.
Change-Id: I964f2e8d351fa5eae9f72b996dda050f5e094f42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2720321
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#858558}


CrOS-Libchrome-Original-Commit: dccec37e8e8c3714192af62cb8eabf4ba961b6b2