Fix a race that leaves FilePathWatcherImpl reference behind
There is a possible race, especially on ChromeOS, that causes
FilePathWatcherImpl::UpdateRecursiveWatches to leave its reference
in `g_inotify_reader`. This CL fixes the issue by making
UpdateRecursiveWatches remove watchers all dirs of `changed_dir`
from `g_inotify_reader` first then add them back in FileEnumerator
loop so that both side have a consistent view of the outstanding
watchers.

To be on the safe side, make InotifyReader store FilePathWatcherImpl
as a task runner and a WeakPtr<> and post the calls instead of a
direct call on a raw pointer. So that even if there is a watcher
leak, posted call would be ignored and the crash would not happen.

Bug: 990004
Change-Id: Ic0d6e6547738bd26d127be5c1b1e66b7664c4bba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2715729
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#859857}


CrOS-Libchrome-Original-Commit: 3879f8aefa554ffdbb11a9c15b3e5eb714c6638b