libchrome_tools: uprev: allow control meta key of origianl commit hash
This allows caller to use a different meta key when appending original
commit hash to commit message.

This is important to not to confuse next generate_filtered_tree run to
build a commit mapping. It can be used to prevent using
recover-new-file-history commits as a parent, leading with more
misleading git blame.

BUG=b:180558819
TEST=manual run with/without args given. check committed commit.

Change-Id: I95e82a7a2fcde49941fb0d16a979a6a4343be7fa
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/external/libchrome/+/2790294
Tested-by: Qijiang Fan <fqj@google.com>
Commit-Queue: Qijiang Fan <fqj@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>