libchrome_tools: merge_main should commit with original author.
merge_main will make author to be the person running the script, because
the intermediate `git merge` commit is authored by the current
developer.

But the script is used to update a working draft of libchrome uprev, to
make parent1=new main tot. Thus we should keep the author of the
original uprev commit.

BUG=None
TEST=merge_main

Change-Id: I9589abd618d02093df08e8abccc6e7a32f392094
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/external/libchrome/+/2919514
Commit-Queue: Qijiang Fan <fqj@google.com>
Tested-by: Qijiang Fan <fqj@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>